### 2021-12-01 #3972 @whedon
[PRE REVIEW]: rmap: An R package to plot and compare tabular data on customizable maps across scenarios and time
**Submitting author:** @zarrarkhan (<a href="http://orcid.org/0000-0002-8147-8553">Zarrar Khan</a>)
**Repository:** <a href="https://github.com/JGCRI/rmap" target ="_blank">https://github.com/JGCRI/rmap</a>
**Version:** v1.0.0
