### 2021-12-01 #3972 @kthyng
Hi @zarrarkhan and thanks for your submission. I am pinging the editorial board for input to make sure this submission is in scope for the journal. Usually we do this for plotting-focused packages to make sure the breadth of the package is large enough. This process will take 1-2 weeks. Thanks for your patience.
