+++ CITATION.md
@@ -1,0 +1,27 @@
+
+<!-------------------------->
+<!-------------------------->
+# Citation
+<!-------------------------->
+<!-------------------------->
+
+Khan, Z., Zhao, M., Wild, T., Vernon, C., 2022. rmap - An R package to plot and compare tabular data on customizable maps across scenarios and time. (In progress) Journal of Open Source Software, DOI: XXXX
+
+
+<!-------------------------->
+<!-------------------------->
+# BibText
+<!-------------------------->
+<!-------------------------->
+
+```
+@article{khan2022rmap,
+  title={gcamextractor: An R package to extract and process GCAM data into standardized tables and units},
+  author = {Zarrar Khan and Mengqi Zhao and Chris Vernon and Thomas Wild},
+  year = {2022},
+  note = {SUBMITTED},
+  url = {https://github.com/JGCRI/rmap},
+  journal={Journal of Open Source Software},
+  doi = {xxx}
+}
+```
