+++ _pkgdown.yml
@@ -16,8 +16,11 @@ navbar:
       href: news/index.html
     - text: "Contribute"
       href: CONTRIBUTE.html
-    - text: cheatsheet
+    - text: Cheatsheet
       href: https://jgcri.github.io/rmap/cheatsheet.pdf
+    - icon: fas fa-quote-left
+      text: Citation
+      href: citation.html
       
 toc:
   depth: 3
