+++ CITATION.md
@@ -5,7 +5,7 @@
 <!-------------------------->
 <!-------------------------->
 
-Khan, Z., Zhao, M., Wild, T., Vernon, C., 2022. rmap - An R package to plot and compare tabular data on customizable maps across scenarios and time. (In progress) Journal of Open Source Software, DOI: XXXX
+Khan, Z., Zhao, M., Wild, T., Vernon, C., 2022. rmap: An R package to plot and compare tabular data on customizable maps across scenarios and time. (In progress) Journal of Open Source Software, DOI: XXXX
 
 
 <!-------------------------->
@@ -16,7 +16,7 @@ Khan, Z., Zhao, M., Wild, T., Vernon, C., 2022. rmap - An R package to plot and 
 
 ```
 @article{khan2022rmap,
-  title={gcamextractor: An R package to extract and process GCAM data into standardized tables and units},
+  title={rmap: An R package to plot and compare tabular data on customizable maps across scenarios and time},
   author = {Zarrar Khan and Mengqi Zhao and Chris Vernon and Thomas Wild},
   year = {2022},
   note = {SUBMITTED},
