+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.0
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2021-12-01T20:36Z
+last_built: 2021-12-02T00:28Z
 
