+++ _pkgdown.yml
@@ -1,5 +1,9 @@
 title: rmap
-  
+
+home:
+  sidebar:
+    structure: [links, license, community, authors, dev]
+
 navbar:
   type: inverse
   left:
