### 2021-12-02 #3972 @zarrarkhan
> @zarrarkhan if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/joss-reviewers) have already agreed to review for JOSS and may be suitable for this submission (please start at the bottom of the list).

Here are suggested reviewers from the list. These reviewers are from the bottom of the list and indicate both R as their preferred language and spatial analysis in their domains:

- fernandomayer
- hknd23
- nickbearman
- ateucher
- tsamsonov
- yongzesong
- ssujit
- jpiburn
- wcjochem
- adithirgis
- kauedesousa
- GISerDaishaoqing
- mikerspencer
- jayrobwilliams
- mikejohnson51
- garretrc
- becarioprecario
- adibender
