### 2021-12-05 #3972 @Kevin-Mattheus-Moerman
@zarrarkhan it would be helpful if you could clearly illustrate how this work has a **direct scientific research application** (beyond visualization), so how it directly influences scientific research. This should make it clear why researchers would feel the need to cite this work. Simple utility tools or visualization tools tend not to be cited since they might not have an effect on the scientific findings. If you could help clarify where this software sits in the spectrum ranging from pure research software to pure visualization, that would be helpful. Are the comparisons only visual or also quantitative?
