### 2021-12-05 #3972 @zarrarkhan
@Kevin-Mattheus-Moerman. Thanks for the question. `rmap's` absolute and percentage difference mapping capabilities while using relatively simple algorithms are not trivial (especially for gridded time series) and definitely have a significant **direct scientific research application** in that they help researches identify changes between scenarios as well as trends across time. All the papers already using `rmap's` capabilities would have cited this paper had it already been published. Additionally as the package is becoming more popular at our institute and partner universities several authors have asked me for how to cite the software. I am hesitant to provide them with a zenodo doi for the software release and would rather provide the final JOSS citation if accepted. The following examples from actual publications highlight the quantitative applications vs pure visualization:

- Khan, Z., Iyer, G., Patel, P., Kim, S., Hejazi, M., Burleyson, C. and Wise, M., 2021. Impacts of long-term temperature change and variability on electricity investments. Nature communications, 12(1), pp.1-12. https://doi.org/10.1038/s41467-021-21785-1
    * Fig 1. `rmap` Pure visualization
    * Fig 4. `rmap` Quantitative spatial analysis of spatial distribution of installed capacity and capital investments across the US in 2100.

- Wild, T.B., Khan, Z., Zhao, M., Suriano, M., Bereslawski, J.L., Roberts, P., Casado, J., Gaviño‐Novillo, M., Clarke, L., Hejazi, M. and Miralles‐Wilhelm, F., 2021. The Implications of Global Change for the Co‐Evolution of Argentina's Integrated Energy‐Water‐Land Systems. Earth's Future, 9(8), p.e2020EF001970. https://doi.org/10.1029/2020EF001970
    * Fig 2. `rmap` Pure visualization
    * Fig 4. `rmap` Quantitative spatial analysis of spatial differences between reference case and multiple Global Climate Model (GCMs) as well as Representative Concentration Pathways (RCPs) to explore implications of climate projections on runoff.
    * Figs 7,8,9: `rmap` Quantitative spatial analysis of the implications of climate change vs policy on landuse and agricultural production. Key to highlighting the implications of these scenarios which would not be possible by comparing pure absolute values across the maps.
    * Fig 10. `rmap` Pure visualization of absolute values to show final conclusion.

- Khan, Z., Wild, T.B., Iyer, G.C., Hejazi, M. and Vernon, C.R., 2021. The future evolution of energy-water-agriculture interconnectivity across the US. Environmental Research Letters. https://doi.org/10.1088/1748-9326/ac046c
    * Fig 5. `rmap` Quantitative spatial analysis of changes in the intersectoral interconnectivity indices across the US from the reference case. In this particular example it was essential to use `rmap's` % difference functionality because we are comparing across distinct parameters (Magnitude Index vs Spread Index) which could not be comparable as absolute values.

- Upcoming research article on implications of socioeconomics and climate change on global water withdrawals which was presented at EGU. The paper is in prep and will be submitted to Nature Scientific Data in the next few months. Preliminary results can be seen at the EGU site: https://presentations.copernicus.org/EGU21/EGU21-903_presentation.pdf
    * Fig. `rmap` Quantitative spatial analysis (global gridded) of the implications of socioeconomics and climate change scenarios by 2100 from the reference scenario (SSP2 no climate in 2015) for global water withdrawals.

In all these papers the difference maps are used to derive the key conclusions of each paper. So I believe `rmap` offers a valuable quantitative capability to make conclusions about the spatial implications of different scenarios and time periods.
