### 2021-12-20 #4015 @whedon
[REVIEW]: rmap: An R package to plot and compare tabular data on customizable maps across scenarios and time
**Submitting author:** @zarrarkhan (<a href="http://orcid.org/0000-0002-8147-8553">Zarrar Khan</a>)
**Repository:** https://github.com/JGCRI/rmap
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.5

- [x] #4016
