### 2021-12-20 #3972 @hugoledoux
I've got 2 reviewers so I'll start the review now, but if one of you wants to the 3rd one let me know I'll add you, first-come-first-served.

@jhollist @mikejohnson51 @jayrobwilliams @GISerDaiShaoqing
