+++ man/mapping_tethys_grid_basin_region_country.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{mapping_tethys_grid_basin_region_country}
+\alias{mapping_tethys_grid_basin_region_country}
+\title{mapping_tethys_grid_basin_region_country}
+\format{
+R table
+}
+\source{
+tethys, gcam
+}
+\usage{
+mapping_tethys_grid_basin_region_country
+}
+\description{
+mapping_tethys_grid_basin_region_country
+}
+\examples{
+\dontrun{
+ library(rmap);
+ mapping_tethys_grid_basin_region_country <- rmap::mapping_tethys_grid_basin_region_country
+}
+}
+\keyword{datasets}
