+++ R/mappings.R
@@ -1,700 +1,0 @@
-#' mappings
-#'
-#' This function has all the mappings used to convert between different categories
-#'
-#' List of Mappings
-#' @param name Default=NULL. Name of assumption object.
-#' \itemize{
-#' \item mapParamQuery}
-#' @keywords assumptions
-#' @return A list of assumptions
-#' @export
-#' @examples
-#' library(rmap)
-#' a<-mappings()
-#' a # will give full list of mappings
-#' @importFrom magrittr %>%
-
-mappings <- function(name=NULL) {
-
-  #------------------------------------------------------------
-  # mapping_tethys_grid_basin_region_country
-  #-----------------------------------------------------------
-
-  mapping_tethys_grid_basin_region_country <- rmap::mapping_tethys_grid_basin_region_country
-
-  #------------------------------------------------------------
-  # Param, query, palette, maps Mapping
-  #-----------------------------------------------------------
-
-  mapParamQuery <- tibble::tribble(
-    ~group, ~param, ~query,~mapPalette,
-    "energy","energyPrimaryByFuelEJ","primary energy consumption by region (direct equivalent)","pal_hot",
-    "energy","energyPrimaryRefLiqProdEJ", "refined liquids production by subsector","pal_hot",
-    "energy","energyFinalConsumBySecEJ", "total final energy by aggregate sector","pal_hot",
-    "energy","energyFinalByFuelBySectorEJ", "Final energy by detailed end-use sector and fuel","pal_hot",
-    "energy","energyFinalSubsecByFuelTranspEJ", "transport final energy by fuel","pal_hot",
-    "energy","energyFinalSubsecByFuelBuildEJ", "building final energy by fuel","pal_hot",
-    "energy","energyFinalSubsecByFuelIndusEJ", "industry final energy by fuel","pal_hot",
-    "energy","energyFinalSubsecBySectorBuildEJ", "building final energy by subsector","pal_hot",
-    "energy","energyFinalConsumByIntlShpAvEJ", "transport final energy by mode and fuel","pal_hot",
-    "energy","energyPrimaryByFuelMTOE", "primary energy consumption by region (direct equivalent)","pal_hot",
-    "energy","energyPrimaryRefLiqProdMTOE", "refined liquids production by subsector","pal_hot",
-    "energy","energyFinalConsumBySecMTOE", "total final energy by aggregate sector","pal_hot",
-    "energy","energyFinalbyFuelMTOE", "Final energy by detailed end-use sector and fuel","pal_hot",
-    "energy","energyFinalSubsecByFuelTranspMTOE", "transport final energy by fuel","pal_hot",
-    "energy","energyFinalSubsecByFuelBuildMTOE", "building final energy by fuel","pal_hot",
-    "energy","energyFinalSubsecByFuelIndusMTOE", "industry final energy by fuel","pal_hot",
-    "energy","energyFinalSubsecBySectorBuildMTOE", "building final energy by subsector","pal_hot",
-    "energy","energyFinalConsumByIntlShpAvMTOE", "transport final energy by mode and fuel","pal_hot",
-    "energy","energyPrimaryByFuelTWh", "primary energy consumption by region (direct equivalent)","pal_hot",
-    "energy","energyPrimaryRefLiqProdTWh", "refined liquids production by subsector","pal_hot",
[…]
