+++ man/mapping_US49.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{mapping_US49}
+\alias{mapping_US49}
+\title{mapping_US49}
+\format{
+R table
+}
+\source{
+rmap
+}
+\usage{
+mapping_US49
+}
+\description{
+mapping_US49
+}
+\examples{
+\dontrun{
+ library(rmap);
+ mapping_US49 <- rmap::mapping_US49
+}
+}
+\keyword{datasets}
