+++ man/mapping_param_query.Rd
@@ -1,0 +1,25 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/data.R
+\docType{data}
+\name{mapping_param_query}
+\alias{mapping_param_query}
+\title{mapping_param_query}
+\format{
+R table
+}
+\source{
+rmap
+}
+\usage{
+mapping_param_query
+}
+\description{
+mapping_param_query
+}
+\examples{
+\dontrun{
+ library(rmap);
+ mapping_param_query <- rmap::mapping_param_query
+}
+}
+\keyword{datasets}
