+++ NAMESPACE
@@ -4,10 +4,8 @@ export(df_to_shape)
 export(map)
 export(map_find_df)
 export(map_plot)
-export(mappings)
 export(printPdfPng)
 export(shape_to_df)
 export(tidy_shape)
 importFrom(jgcricolors,jgcricol)
-importFrom(magrittr,"%>%")
 importFrom(rlang,":=")
