+++ pkgdown.yml
@@ -1,7 +1,7 @@
 pandoc: 2.7.3
-pkgdown: 2.0.0
+pkgdown: 2.0.1
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2021-12-02T14:07Z
+last_built: 2021-12-22T17:32Z
 
