+++ R/df_to_shape.R
@@ -7,6 +7,7 @@
 #' @keywords map, find
 #' @return SpatialPolygonDataFrame
 #' @export
+#' @importFrom magrittr %>%
 #' @examples
 #' library(rmap)
 #' df = rmap::mapCountriesdf
