+++ .github/workflows/build.yml
@@ -15,7 +15,7 @@ jobs:
       matrix:
         config:
         - { os: windows-latest, r: 'devel'}
-        - { os: macOS-latest, r: 'devel'}
+        - { os: macos-10.15, r: 'devel'}
         - { os: ubuntu-latest, r: 'devel', cran: "https://demo.rstudiopm.com/all/__linux__/xenial/latest"}
     
     env:
