+++ NAMESPACE
@@ -8,4 +8,5 @@ export(printPdfPng)
 export(shape_to_df)
 export(tidy_shape)
 importFrom(jgcricolors,jgcricol)
+importFrom(magrittr,"%>%")
 importFrom(rlang,":=")
