+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.1
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2021-12-22T17:32Z
+last_built: 2021-12-22T18:12Z
 
