### 2021-12-29 98468b0
Updating comments from https://github.com/openjournals/joss-reviews/issues/4015#issuecomment-1002515157
