+++ paper/paper.bib
@@ -40,13 +40,13 @@
 	title = {tmap: {Thematic} {Maps} in {R}},
 	volume = {84},
 	shorttitle = {tmap},
+	doi = {10.18637/jss.v084.i06},
 	url = {https://github.com/r-tmap/tmap},
 	number = {1},
 	journal = {Journal of Statistical Software},
 	author = {Tennekes, Martijn},
 	year = {2018},
-	pages = {1--39},
-	file = {Full Text:C\:\\Users\\khan404\\Zotero\\storage\\FCZWIXJE\\Tennekes - 2018 - tmap Thematic Maps in R.pdf:application/pdf;Snapshot:C\:\\Users\\khan404\\Zotero\\storage\\FHBEF22A\\v084i06.html:text/html},
+	pages = {1--39}
 }
 
 @article{giraud_cartography_2016,
@@ -165,7 +165,7 @@
 	title = {The {Implications} of {Global} {Change} for the {Co}-{Evolution} of {Argentina}'s {Integrated} {Energy}-{Water}-{Land} {Systems}},
 	volume = {9},
 	number = {8},
-	doi = {0.1029/2020EF001970},
+	doi = {10.1029/2020EF001970},
 	journal = {Earth's Future},
 	author = {Wild, Thomas B. and Khan, Zarrar and Zhao, Mengqi and Suriano, Micaela and Bereslawski, Julia L. and Roberts, Paula and Casado, Jose and Gaviño-Novillo, Marcelo and Clarke, Leon and Hejazi, Mohamad},
 	year = {2021},
