### 2021-12-29 #4015 @zarrarkhan
Hi @CamilleMorlighem,

Thanks for the very useful feedback and comments. Here are the responses to your points:

1. We fixed the typo on line 17 and also fixed the Missing and incorrect DOIs. The updated document can be found here: http://res.cloudinary.com/hkvhhndao/image/upload/v1640793026/zeldl1tq79mt9xnk82us.pdf or in the repo: https://github.com/JGCRI/rmap/tree/main/paper
2. We updated the BSD_2_clause license file. Now when you click on the repo LICENSE file (https://github.com/JGCRI/rmap/blob/main/LICENSE) it should show the proper BSD_2_clause license.
3. We added a [support tab](https://jgcri.github.io/rmap/SUPPORT.html) in the user guide and setup a [discussions](https://github.com/JGCRI/rmap/discussions) page for rmap in the main repo. If users do not want to submit an issue then they can start a discussion instead on the [discussions](https://github.com/JGCRI/rmap/discussions) page. We the developers can respond there as well as have community members respond there as well.
