### 2021-12-31 #4015 @CamilleMorlighem
Hi @zarrarkhan,

Thank you for your responses !

I have tested the documentation and functionality of the package. The documentation is very clear and structured and very user friendly! I just have the following comments for improvement:
1. The information link (https://confluence.pnnl.gov/confluence/display/JGCRI/GCAM+Shape+Files) in the help section of the maps `mapGCAMBasins` and `mapGCAMReg32` is not working for me (I reach an error page).
2. I could not find a statement of need in the user guide or in the documentation on the repository.
3. I couldn't find either a list of dependencies, but when installing it proposed me to install/update required packages so maybe that's enough?
4. I found some typos in the user guide:
- In the section "Subset existing shape", I think you meant `unique(shapeSubset@data$subRegion)` and not `unique(shapeSubset@data$states)`
- In section "Read data from a shapefile", I think you meant `rgdal::readOGR(dsn=exampleFolder,layer=exampleFile,use_iconv=T,encoding='UTF-8')` instead of `rgdal::readOGR(dsn=examplePolyFolder,layer=examplePolyFile,use_iconv=T,encoding='UTF-8')`.
- In subsection "Gridded Data Multi-Year Diff", both code snippets return NULL:
                                   `mapx$map_param_xDiffAbs_2015_1990_KMEANS`
				   `mapx$map_param_xDiffPrcnt_2015_1990_KMEANS`
- Same with the following code snippets in section "Gridded Data Multi-Scenario Diff":
                                   `mapx$map_param_DiffAbs_scenario3_scenario1_KMEANS`
				   `mapx$map_param_DiffPrcnt_scenario3_scenario1_KMEANS`

About the functionality, everything worked fine and I don't have any performance claims. I only have the following comments:
1. I did not get the effect of the argument `scaleRangeDiffAbs` (or `scaleRangeDiffPrcnt`) in the `map` function . I could not see what changed in the results when changing the values for these arguments or when simply removing them.
2. I thought that one very useful function, that is not described in the documentation, is the `map_find_df` function to find a suitable map for your dataframe. I think it might be very useful when you don't know the package and want to use it for a specific purpose, so maybe it is worth highlighting this function in the user-guide?

Feel free to ignore any comment if you think that's not relevant!
