### 2022-01-03 51c15ae
Updating rmap in response to https://github.com/openjournals/joss-reviews/issues/4015#issuecomment-1003394083
