### 2022-01-03 #4015 @zarrarkhan
@CamilleMorlighem Thanks for the detailed review and very helpful comments. Addressed as follows:

**Improvements**
1. All data sources referring to the restricted (https://confluence.pnnl.gov/confluence/display/JGCRI/GCAM+Shape+Files) have been updated to point to the publicly available zenodo link: https://zenodo.org/record/4688451#.YdMNTmjMJPY
2. We added statement of need to the repo and user guide landing pages based on the text in the paper.
- https://github.com/JGCRI/rmap/#statement-of-need
- https://jgcri.github.io/rmap/#statement-of-need
3. Dependencies can be found in the https://github.com/JGCRI/rmap/blob/main/DESCRIPTION file and those mentioned in the imports list are required and will be automatically installed when users install the package so it shouldn't be something a user needs to worry about outside the package.
4. Typos in user guide fixed (Thanks for catching these!):
* In the section "Subset existing shape", unique(shapeSubset@data$states) siwtched to unique(shapeSubset@data$subRegion).
* In section "Read data from a shapefile", I think you meant
rgdal::readOGR(dsn=examplePolyFolder,layer=examplePolyFile,use_iconv=T,encoding='UTF-8') changed to rgdal::readOGR(dsn=exampleFolder,layer=exampleFile,use_iconv=T,encoding='UTF-8')
* In subsection "Gridded Data Multi-Year Diff", code snippets adjusted to return appropriate maps. The code snippet was out of date.
* Same with the following code snippets in section "Gridded Data Multi-Scenario Diff" code snippets adjusted to return appropriate maps. The code snippet was out of date.

**Functionality**
1. Thanks for noting this. The code needed to be updated to implement the changes. I think with the updated changes (now that it is actually working) you can see the usefulness of this feature (i.e. highlighting different ranges for each type of map as needed). See here: https://jgcri.github.io/rmap/articles/vignette_map.html#scale-range
2. Added section on map_find_df to user_guide here: https://jgcri.github.io/rmap/articles/vignette_map.html#map-find
