+++ vignettes/vignette_map.Rmd
@@ -288,7 +288,15 @@ rmap::map(map_chosen) # This can be plotted using rmap if desired
 
 <!-------------------------->
 <!-------------------------->
-# Layers {#layers}
+# Format (Themes, layers, colors, zoom)
+<!-------------------------->
+<!-------------------------->
+<p align="center"> <img src="vignetteFigs/divider.png"></p>
+
+
+<!-------------------------->
+<!-------------------------->
+## Layers {#layers}
 <!-------------------------->
 <!-------------------------->
 <p align="center"> <img src="vignetteFigs/divider.png"></p>
@@ -326,7 +334,7 @@ rmap::map(data,
 
 <!-------------------------->
 <!-------------------------->
-# Layer Colors {#layercolors}
+## Layer Colors {#layercolors}
 <!-------------------------->
 <!-------------------------->
 <p align="center"> <img src="vignetteFigs/divider.png"></p>
@@ -374,7 +382,7 @@ rmap::map(data,
 
 <!-------------------------->
 <!-------------------------->
-# Labels {#labels}
+## Labels {#labels}
 <!-------------------------->
 <!-------------------------->
 <p align="center"> <img src="vignetteFigs/divider.png"></p>
@@ -453,7 +461,7 @@ rmap::map(data,
 
 <!-------------------------->
 <!-------------------------->
-# Crop {#crop}
+## Crop {#crop}
 <!-------------------------->
 <!-------------------------->
 <p align="center"> <img src="vignetteFigs/divider.png"></p>
@@ -498,7 +506,7 @@ rmap::map(data,
 
 <!-------------------------->
 <!-------------------------->
-# Zoom {#zoom}
+## Zoom {#zoom}
 <!-------------------------->
 <!-------------------------->
 <p align="center"> <img src="vignetteFigs/divider.png"></p>
@@ -546,7 +554,7 @@ cowplot::plot_grid(m0[[1]],m1[[1]], m2[[1]],m3[[1]], m4[[1]],
 
 <!-------------------------->
 <!-------------------------->
-# Background {#background}
+## Background {#background}
 <!-------------------------->
 <!-------------------------->
 <p align="center"> <img src="vignetteFigs/divider.png"></p>
@@ -585,7 +593,7 @@ rmap::map(data,
 
 <!-------------------------->
 <!-------------------------->
-# Themes {#themes}
+## Themes {#themes}
 <!-------------------------->
 <!-------------------------->
 <p align="center"> <img src="vignetteFigs/divider.png"></p>
@@ -813,7 +821,6 @@ my_custom_map@data <- my_custom_map@data %>% dplyr::mutate(subRegion="CHOOSE_APP
 data <- data.frame(subRegion=unique(my_custom_map@data$subRegion),
                    value = 100*runif(length(unique(my_custom_map@data$subRegion))));
 head(data)
-
 # And then it can be plotted using rmap and any of the built-in maps as underLayers.
 rmap::map(data,
           shape=my_custom_map,
@@ -1016,8 +1023,8 @@ mapx <- rmap::map(data = data,
           xDiff = c(2010,2015),
           show=F, save = T)
 
-mapx$map_param_xDiffAbs_2015_1990_KMEANS
-mapx$map_param_xDiffPrcnt_2015_1990_KMEANS
+mapx$map_param_KMEANS_xDiffAbs
+mapx$map_param_KMEANS_xDiffPrcnt
 
 ```
 
@@ -1311,12 +1318,12 @@ mapx$map_param_KMEANS_DiffPrcnt
[…]
