+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.1
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2022-01-03T19:39Z
+last_built: 2022-01-03T20:36Z
 
