### 2022-01-04 8cafce6
Updating rmap in response to https://github.com/openjournals/joss-reviews/issues/4015#issuecomment-1004832075
