### 2022-01-04 #4015 @zarrarkhan
Hi @maczokni.
Thanks for your review and comments. Responses below:

- Column names (subRegion and Value). Have highlighted how users can identify their own column names in rmap using the `subRegCol` and `valueCol` arguments.
