### 2022-01-04 d0303b2
Updating rmap issue with spatialPolygonDataFrame
