+++ vignettes/vignette_map.Rmd
@@ -89,9 +89,9 @@ Users can also assign other column names to `subRegion` and `value` if desired u
 library(rmap);
 data = data.frame(my_column_1 = c("CA","FL","ID","MO","TX","WY"),
                   my_column_2 = c(5,10,15,34,2,7))
-map(data,
-    subRegCol = "my_column_1",
-    valueCol = "my_column_2")
+rmap::map(data,
+          subRegCol = "my_column_1",
+          valueCol = "my_column_2")
 
 ```
 
