### 2022-01-04 #4015 @zarrarkhan
Hi @maczokni.

Thanks for your review and comments. Please see our responses below:

```
# Update rmap with latest changes
devtools::install_github("JGCRI/rmap")
library(rmap)
```

1. Required Column names (subRegion and Value). We have highlighted how users can identify their own column names in rmap using the subRegCol and valueCol arguments. (see end of this section: https://jgcri.github.io/rmap/articles/vignette_map.html#inputs)
2. Example data. Thanks for trying rmap with your example data. So we updated a few things and have some comments on this:
* `ncovr <- geodaData::ncov` gives an sf object with geometries. rmaps was getting unable to process this because it was not a dataframe. Code updated to convert an sf object to a pure dataframe.
* mydata <- ncovr %>% select(NAME, HR60) results in a dataframe with a list of counties and corresponding data. If you sort by NAME you will see multiple entries for the same subRegion. `rmap` (or any other mapping package for that matter) would not know how to deal with this repeated data. We could filter for the first entries or something like that but I think it is better to throw an error.
```ncovr %>%
  dplyr::select(NAME,HR60) %>%
  as.data.frame() %>%
  dplyr::arrange(NAME) %>%
  head(10)
```

3. With the updates you can run your example as follows. Note that in general, county data is problematic because there are multiple counties with the same name in different states. In order to be more specific you can append your county names with an underscore and relevant State initials.
```
library(dplyr)
library(geodaData)
library(devtools)
library(rmap)

ncovr <- geodaData::ncovr

# Subset first 10 rows to avoid repeated subRegions
mydata <- ncovr %>%
  dplyr::select(NAME,HR60) %>%
  head(10); mydata

# Will give you the relevant plot but multiple counties
# See how you can define your own columns as arguments
rmap::map(mydata,
          subRegCol = "NAME",
          valueCol = "HR60")

# Here you can see all the multiple counties labelled.
# rmap appends the State for you as it recognizes this.
# You can also see some of the other features of rmap here
# such as underLayer, zoom, labels etc.
rmap::map(mydata,
          subRegCol="NAME",
          valueCol="HR60",
          labels = T,
          labelSize = 3,
          labelRepel = T,
          underLayer = rmap::mapUS49,
          zoom=-2)
```

Please let us know if this helps clarify some of the issues you were having and if you have additional concerns/comments.
