+++ inst/extras/saveDataFiles.R
@@ -128,6 +128,20 @@ if(redoMaps){
   mapStates@data <- mapStates@data %>%
     dplyr::mutate(name="mapStates")
   use_data(mapStates, version=3, overwrite=T)
+
+  mapx <- rmap::mapStates
+  mapx <- mapx[!is.na(mapx@data$subRegion),]
+  mapx@data <- mapx@data%>%droplevels()
+  mapx@data %>% filter(is.na(subRegion))
+  nrow(mapx)
+  mapx@data <- mapx@data %>%
+    dplyr::mutate(subRegion = if_else(is.na(subRegion),subRegionAlt,subRegion)) %>%
+    dplyr::filter(!is.na(subRegion));
+  mapx@data %>% filter(is.na(subRegion))
+  nrow(mapx)
+  mapStates <- mapx
+  use_data(mapStates, version=3, overwrite=T)
+
 }
 
 #-----------------
@@ -772,6 +786,27 @@ if(redoMaps){
   mapUS52County@data <- mapUS52County@data %>%
     dplyr::mutate(name="mapUS52County")
   use_data(mapUS52County, version=3, overwrite=T)
+
+  # Create FIPS
+  mapx <- rmap::mapUS52County
+  mapx@data <- mapx@data  %>%
+    dplyr::mutate(subRegion_cc = paste0(subRegion,COUNTYCODE))
+  subRegions_duplicated <- (mapx@data$subRegion)[(mapx@data$subRegion)%>%duplicated()]; subRegions_duplicated
+  countycodes_remove <- mapx@data %>%
+    dplyr::select(subRegion,subRegionAlt,COUNTYCODE, area_sqkm, subRegion_cc) %>%
+    dplyr::filter(subRegion %in% subRegions_duplicated)%>%
+    dplyr::group_by(subRegionAlt) %>%
+    dplyr::mutate(max_area = max(area_sqkm)) %>%
+    dplyr::filter(area_sqkm != max_area); countycodes_remove
+
+  mapx@data <- mapx@data %>%
+    dplyr::mutate(
+      subRegion = if_else(subRegion_cc %in% countycodes_remove$subRegion_cc,paste0(subRegion,"_city"),subRegion),
+      FIPS = paste0(STATEFP,COUNTYCODE))
+  mapUS52County <- mapx
+  subRegions_duplicated <- (mapx@data$subRegion)[(mapx@data$subRegion)%>%duplicated()]; subRegions_duplicated
+  use_data(mapUS52County, version=3, overwrite=T)
+
 }
 
 # US 52 Counties with Alaska (AK), Hawaii (HI) and Puerto Rico (PR) shrunken and shifted
@@ -794,19 +829,19 @@ if(redoMaps){
   alaska <- elide(alaska, scale=max(apply(bbox(alaska), 1, diff)) / 2.3)
   alaska <- elide(alaska, shift=c(-2100000, -2500000))
   proj4string(alaska) <- proj4string(us_aea)
-  rmap::rmap::map(alaska)
+  rmap::map(alaska)
   # extract, then rotate & shift hawaii
   hawaii <- us_aea[us_aea$STATENAME=="Hawaii",]
   hawaii <- elide(hawaii, rotate=-35)
   hawaii <- elide(hawaii, shift=c(5400000, -1400000))
   proj4string(hawaii) <- proj4string(us_aea)
-  rmap::rmap::map(hawaii)
+  rmap::map(hawaii)
   # extract, then rotate & shift Puerto Rico
   pr <- us_aea[us_aea$STATENAME=="Puerto Rico",]
   #pr <- elide(pr, rotate=-35)
   pr <- elide(pr, shift=c(-2500000,0))
   proj4string(pr) <- proj4string(us_aea);
-  rmap::rmap::map(pr)
[…]
