+++ pkgdown.yml
@@ -1,7 +1,7 @@
 pandoc: 2.7.3
-pkgdown: 2.0.1
+pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2022-01-04T23:02Z
+last_built: 2022-01-22T13:12Z
 
