### 2022-01-23 #4015 @zarrarkhan
@maczokni
Thanks for your comments and notes. We have addressed your comments as follows:

- US County duplicates:
rmap::mapUS49County@data (which is from the official census.gov webpage: https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html) had all unique rows, even though as you found, six subRegions had duplicate names in spite of being unique entries (which we had missed). Initially, since there were so many counties with the same name in different states we had attempted to address this by appending State and County names together in the subRegion column as shown below. As you can see each subRegion has a unique STATEFP & COUNTYCODE combination even though the subRegion names may have been duplicated in these 6 particular cases (Additionally COUNTYCODEs are also repeated in different states):
```
        subRegion       subRegionAlt STATEFP COUNTYCODE
1       Baltimore_MD    Baltimore    24        005
2       Baltimore_MD    Baltimore    24        510
3       Fairfax_VA      Fairfax      51        059
4       Fairfax_VA      Fairfax      51        600
5       Franklin_VA     Franklin     51        067
6       Franklin_VA     Franklin     51        620
7       Richmond_VA     Richmond     51        159
8       Richmond_VA     Richmond     51        760
9       Roanoke_VA      Roanoke      51        161
10      Roanoke_VA      Roanoke      51        770
11      St. Louis_MO    St. Louis    29        189
12      St. Louis_MO    St. Louis    29        510
```
- Duplicate subRegions in rmap::mapUS49County
There were six subRegions in rmap::mapUS49County which had duplicated subRegion names as shown above. These occurred because each of these subRegions are classified as both **cities** as well as **counties** and are allocated their own COUNTYCODES in spite of being cities. For each of these cities we also have a county with the same name in the same state. **For each city we have appended their names with `_city` so that we have completely unique subRegions as well as added FIPS as per your suggestion**. Updated table outputs (selected columns) now show:
```
           subRegion subRegionAlt STATEFP COUNTYCODE  FIPS
1       Baltimore_MD    Baltimore      24        005 24005
2  Baltimore_MD_city    Baltimore      24        510 24510
3         Fairfax_VA      Fairfax      51        059 51059
4    Fairfax_VA_city      Fairfax      51        600 51600
5        Franklin_VA     Franklin      51        067 51067
6   Franklin_VA_city     Franklin      51        620 51620
7        Richmond_VA     Richmond      51        159 51159
8   Richmond_VA_city     Richmond      51        760 51760
9         Roanoke_VA      Roanoke      51        161 51161
10   Roanoke_VA_city      Roanoke      51        770 51770
11      St. Louis_MO    St. Louis      29        189 29189
12 St. Louis_MO_city    St. Louis      29        510 29510
```
- 3 missing FIPS when joining with `ncovr` dataset
The `not_joined` dataframe in your example above refers to 3 counties which are out of date in the `ncovr` dataset. These three subRegions from the `not_joined` datatable are shown below. References to when their names were changed are listed after the code chunk:
```
not_joined %>% dplyr::select(NAME,STATE_NAME,STATE_FIPS,CNTY_FIPS,FIPS)
           NAME   STATE_NAME STATE_FIPS CNTY_FIPS  FIPS
1       Shannon South Dakota         46       113 46113
2 Clifton Forge     Virginia         51       560 51560
3          Dade      Florida         12       025 12025
```
(Source: https://www.ddorn.net/data/FIPS_County_Code_Changes.pdf)
1. Shannon: South Dakota, 1983: Washabaugh county (FIPS 46131) merges into
Jackson county (FIPS 46071).
2.  Clifton Forge: Virginia, 2001: The independent city of Clifton Forge (FIPS 51560) merges
into Alleghany county (FIPS 51005).
3. Dade: Florida, 1997: Dade county (FIPS 12025) is renamed as Miami-Dade
county (FIPS 12086).
- Broader concern: Each of our datasets provides the source for the underlying data which can be compared to user data if needed. For example the county data is from "https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html". The datasets also contain additional columns such as FIPS and other details that were part of the original datasets to ensure unique entries. Some degree of data cleaning and checks is expected with user data.
```
# Checking the source of rmap maps
rmap::mapUS49County@data$source%>%unique()
# "https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html"
rmap::mapCountries@data$source%>%unique()
# "https://www.naturalearthdata.com/downloads/"
```

Please let us know if this addresses some of your concerns or if you have additional comments.
