### 2022-02-18 2a303ad
Updating cheatsheet
