+++ .github/workflows/docs.yaml
@@ -47,3 +47,18 @@ jobs:
           git config --local user.name "$GITHUB_ACTOR"
           git config --local user.email "$GITHUB_ACTOR@users.noreply.github.com"
           Rscript -e 'pkgdown::deploy_to_branch(new_process = FALSE)'
+          
+      - name: Track file
+        run: |
+          git add .github/r-depends.rds
+          git commit -m "track file"
+      
+      - name: Rename file
+        run: |
+          git mv csheet.pdf cheatsheet.pdf
+      
+      - name: Auto commit
+        uses: stefanzweifel/git-auto-commit-action@v4
+        with:
+          branch: gh-pages
+          push_options: --force
