+++ pkgdown.yml
@@ -1,7 +1,7 @@
-pandoc: 2.7.3
+pandoc: '2.5'
 pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2022-02-18T03:17Z
+last_built: 2022-02-18T15:56Z
 
