+++ R/map.R
@@ -299,7 +299,7 @@ map <- function(data = NULL,
   }
 
   # Make sure data is a dataframe if not a spatialPolygonDataframe
-  if(!grepl("SpatialPolygonsDataFrame",class(data),ignore.case = T)){
+  if(!any(grepl("SpatialPolygonsDataFrame",class(data),ignore.case = T))){
   data <- tibble::as_tibble(data)}
 
   # Rename SubRegCol
@@ -874,7 +874,7 @@ map <- function(data = NULL,
 
                     tbl_temp2 <-dataTblDiffb%>%
                       dplyr::filter(scenario==scen_i)%>%
-                      dplyr::mutate(!!paste(scen_i,"_xDiffPrcnt_",xRef_i,sep=""):=((!!as.name(x_i)-!!as.name(xRef_i))*100/!!as.name(x_i)),
+                      dplyr::mutate(!!paste(scen_i,"_xDiffPrcnt_",xRef_i,sep=""):=((!!as.name(x_i)-!!as.name(xRef_i))*100/!!as.name(xRef_i)),
                                     palette=paletteDiff)%>%
                       dplyr::select(-!!as.character(xDiff_i),-!!as.character(xRef_i))
                     tbl_temp2<-tbl_temp2%>%
@@ -1584,6 +1584,40 @@ map <- function(data = NULL,
                     mapsReturn[[return_i]];
                   names(mapsReturn)[return_i] <- paste("map_",param_i,"_",fileNameTag,nameAppend,sep="");
                   return_i = return_i + 1
+
+                  # color=color; lwd=lwd; legendType=legendType; save=save;  show=show; shape = shapex; overLayer=overLayer; overLayerColor=overLayerColor;
+                  # overLayerFill = overLayerFill; overLayerLwd = overLayerLwd;
+                  # overLayerAlpha = overLayerAlpha; underLayerColor=underLayerColor;
+                  # underLayerFill = underLayerFill; underLayerLwd = underLayerLwd;
+                  # underLayerAlpha = underLayerAlpha; background=background; zoom=zoom;
+                  # zoomx = zoomx; zoomy=zoomy; asp=asp; legendShow=legendShow;
+                  # crop = crop; crop_to_underLayer = crop_to_underLayer; crop_to_overLayer = crop_to_overLayer; transparent=transparent;
+                  # alpha = alpha; size=max(1,(size+(colm+rowm)*3 - 12)); ncol=ncol;
+                  # showNA=showNA; colorNA=colorNA;
+                  # labelColor=labelColor;
+                  # labelSize=labelSize;
+                  # labelAlpha=labelAlpha;
+                  # labelFill=labelFill;
+                  # labelBorderSize=labelBorderSize;
+                  # theme = theme; legendTitle=legendTitle;
+                  # legendDigitsOverride=legendDigitsOverride;
+                  # numeric2Cat_list=numeric2Cat_list; catParam = param_i;
+                  # underLayer=underLayer;
+                  # data=datax;
+                  # legendBreaksn=legendBreaksn;
+                  # legendDigits = legendDigits;
+                  # palette = palette;
+                  # width=width*max(1,colm/1);
+                  # height=height*max(1,rowm/1);
+                  # pdfpng = pdfpng; legendSingleColor = legendSingleColor; legendSingleValue =  legendSingleValue;
+                  # labels=labels; labelRepel=labelRepel; underLayerLabels=underLayerLabels; overLayerLabels=overLayerLabels;
+                  # legendBreaks = legendBreaksx;
+                  # fillColumn = "value"; shapeColumn = shapeColumn;
+                  # col = multiFacetColsx;
+                  # row = multiFacetRowsx;
[…]
