+++ DESCRIPTION
@@ -34,7 +34,6 @@ Imports:
 	ggplot2,
 	ggrepel,
 	maptools,
-	jgcricolors,
 	cowplot,
     markdown,
 	broom,

[…]
