### 2022-02-28 048091a
Updating rmap to remove NA subregions
