### 2022-02-28 #4015 @zarrarkhan
@maczokni

Thank you for your response. Here is a working example with your dataset after removing out-of-date FIPS as we discussed above. I tried your counter-examples and ran into some issues for which I have shared more details below.

**Working Example with your data**
```
data("ncovr")
mydata <- ncovr %>% select(NAME, STATE_NAME, FIPS, HR60) %>% st_drop_geometry() %>%
  dplyr::left_join(rmap::mapUS52County@data, by="FIPS") %>%
  dplyr::mutate(value=HR60) %>%
  dplyr::filter(!is.na(subRegion)) # Remove non-existent FIPS 51560 (Clifton Forge), FIPS 12025 (Dade)
rmap::map(data=mydata)

```
**Missing FIPS in ncovr data**
You are using a specific dataset "ncovr" which has FIPS for counties that do not exist any more as we described in [#4015](https://github.com/openjournals/joss-reviews/issues/4015#issuecomment-1019419000):
(Source: https://www.ddorn.net/data/FIPS_County_Code_Changes.pdf)
- Shannon: South Dakota, 1983: Washabaugh county (FIPS 46131) merges into
- Jackson county (FIPS 46071).
- Clifton Forge: Virginia, 2001: The independent city of Clifton Forge (FIPS 51560) merges
- into Alleghany county (FIPS 51005).
- Dade: Florida, 1997: Dade county (FIPS 12025) is renamed as Miami-Dade
- county (FIPS 12086).

**Your Counter Example 1**
```
download.file("https://www2.census.gov/geo/tiger/GENZ2018/kml/cb_2018_us_county_20m.zip","counties.zip")
unzip("counties.zip")
counties_geojson <- sf::st_read("cb_2018_us_county_20m.kml")
mapthis <- dplyr::left_join(counties_geojson, mydata, by = c("GEOID" = "FIPS"))
tmap::qtm(mapthis, fill = "HR60")
```
```> mapthis <- dplyr::left_join(counties_geojson, mydata, by = c("GEOID" = "FIPS"))
Error: Join columns must be present in data.
x Problem with `GEOID`.
```

**Your Counter Example 2**
```
tg_counties <- tigris::counties()
mapthis <- dplyr::left_join(tg_counties, mydata, by = c("GEOID" = "FIPS"))
tmap::qtm(mapthis, fill = "HR60")
```
- My R does not give errors but it stalls and is unable to display the final map.
- `rmap` produces the map for this data almost instantly and does not require any downloads

**Updates made and additional comments**
- We have modified the code to check for NA subRegions and so you do not need to filter your data anymore.

**Working Example with your data after updates**
```
# devtools::install_github("JGCRI/rmap") # To update to latest version
library(rmap)
library(geodaData)
library(dplyr)
library(sf)

data("ncovr")
mydata <- ncovr %>% select(NAME, STATE_NAME, FIPS, HR60) %>% st_drop_geometry() %>%
  dplyr::left_join(rmap::mapUS52County@data, by="FIPS") %>%
  dplyr::mutate(value=HR60)
rmap::map(data=mydata)
```
Please let us know if the example above is working for you and if you any have additional comments or issues.
