### 2022-03-01 #4015 @maczokni
Hiya I'll give this a go later thanks but it's still just one example of the counties and it still requires the user to do data cleaning, so still not sure about the target audience being non GIS users.

However in the meantime I wanted to share two concerns with the code:

1 -  the package functions producing character vectors are  created using `print()` rather than messages using `message()`? It's going to be an issue for users because it defeats things like suppressMessages() and knitr chunk options when the package is used in a Rmarkdown document. It would be even better if you could use `rlang::inform()` because it has nice formatting for their long messages while still playing nicely with functions that expect `base::message()`.

2 - you should consider adopting sf for the internal workings rather than sp. Roger Bivand (who wrote the sp package in the first place) recommends at https://doi.org/ghnwg3 that new packages use sf rather than sp and says that packages that continue to depend on SP will have to "hope" it continues to be maintained. I also note that `rmap` depends on `rgdal` and `rgeos`, both of which will be retired by the end of 2023 (as noted in bold type at the top of the NEWS.md pages of both packages) and neither of which are required if you use sf instead.
