### 2022-03-04 da9c415
Updating rmap to remove print statements and dependencies on rgeos and rgdal
