+++ R/map.R
@@ -256,7 +256,7 @@ map <- function(data = NULL,
   # crop_to_underLayer = F
   # crop_to_overLayer = F
 
-  print("Starting map...")
+  rlang::inform("Starting map...")
 
   #.................-
   # Initialize variables
@@ -503,7 +503,7 @@ map <- function(data = NULL,
     # Return Data
     # .................
 
-    print("map run completed.")
+    rlang::inform("map run completed.")
 
     invisible(mapsReturn)
 
@@ -669,9 +669,7 @@ map <- function(data = NULL,
   # Remove NA's & Keep only Unique Values
   if(!is.null(dataTbl)){
     if(nrow(dataTbl)>0){
-    #print("Removing NA's and keeping only unique values in dataTbl...")
     dataTbl<-dataTbl%>%dplyr::filter(!is.na(value))%>%dplyr::mutate(value = signif(value,10))%>%dplyr::ungroup()%>%dplyr::distinct()
-    #print("Complete.")
     }
   }
 
@@ -706,22 +704,22 @@ map <- function(data = NULL,
     if(!is.null(scenRef)){
 
       if(!any(scenRef %in% scenarios)){
-        print(paste("scenRef chosen: ", scenRef, " is not in any of the available scenarios: ",sep=""))
-        print(paste(scenarios,collapse=", "))
-        print(paste("Setting scenRef to first scenario: ", scenarios[1],".",sep=""))
+        rlang::inform(paste0("scenRef chosen: ", scenRef, " is not in any of the available scenarios: "))
+        rlang::inform(paste(scenarios,collapse=", "))
+        rlang::inform(paste0("Setting scenRef to first scenario: ", scenarios[1],"."))
         scenRef <- scenarios[1]
       }
 
 
       if(is.null(scenDiff)){
         scenDiff <- scenarios[!scenarios %in% scenRef]
-        print(paste("Running difference against all available scenarios:",sep=""))
-        print(paste(scenDiff,collapse=", "))
+        rlang::inform(paste0("Running difference against all available scenarios:"))
+        rlang::inform(paste(scenDiff,collapse=", "))
       }else{
         if(!any(scenDiff %in% scenarios)){
-          print(paste("None of the scenDiff are in any of the available scenarios: "))
-          print(paste(scenarios[!scenarios %in% scenRef],collapse=", "))
-          print(paste("Skipping Diff.",sep=""))
+          rlang::inform(paste("None of the scenDiff are in any of the available scenarios: "))
+          rlang::inform(paste(scenarios[!scenarios %in% scenRef],collapse=", "))
+          rlang::inform(paste0("Skipping Diff."))
         }
       }
 
@@ -745,9 +743,9 @@ map <- function(data = NULL,
       if(length(unique(dataTblDiffa$scenario))>1){
 
         if(scenRef_i %in% unique(dataTblDiffa$scenario)){
-          print(paste("Ref scenario chosen for param: ", param_i, " is ", paste(scenRef_i,collapse=", "),sep=""))
+          rlang::inform(paste("Ref scenario chosen for param: ", param_i, " is ", paste(scenRef_i,collapse=", "),sep=""))
         if(any(scenDiff_i %in% unique(dataTblDiffa$scenario))){
-          print(paste("Diff scenarios chosen for param: ", param_i, " are ",
+          rlang::inform(paste("Diff scenarios chosen for param: ", param_i, " are ",
                       paste(scenDiff_i[scenDiff_i %in% unique(dataTblDiffa$scenario)],collapse=", "),sep=""))}
 
 
@@ -809,26 +807,26 @@ map <- function(data = NULL,
     if(!is.null(xRef)){
 
       if(!any(xRef %in% xRange)){
[…]
