+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2022-02-28T17:41Z
+last_built: 2022-03-04T20:20Z
 
