### 2022-03-04 #4015 @zarrarkhan
@maczokni

Thank you for the helpful comments and suggestions. We have addressed these in commit [da9c41574820679fae9e0cb4871cf014edf2213c](https://github.com/JGCRI/rmap/commit/da9c41574820679fae9e0cb4871cf014edf2213c) with responses below:

- **Comment on Target Audience:** Sorry if this wasn't clear but by targeting non-GIS users we did not mean that this would not require data cleaning or processing. `rmap` provides extensive examples for users to build maps of common geographical boundaries (countries, states, basins, counties etc.) with the user requiring a basic table with a minimum of two columns ("subRegion" and "value") as you have seen. This still means the user has to prepare their datasets into those basic tables and may run into incompatibility issues if using region names which don't exist in rmap's databases.
- **Print:** We have replaced all print statements with `rlang::inform`
- **sp, rgdal, rgeos dependencies:** We have removed the dependency on rgdal and rgeos. `rmap` still depends on `sp` as it builds all of its maps from the SpatialPolygonDataFrame class. We agree that this should be updated to `sf` and will do so in v2.0 of the model as the package will need extensive testing for these replacements.
