### 2022-03-05 #4015 @maczokni
I'm really not trying to be difficult, but I believe the package requires extensive testing anyway. I just tried again with a different data set to map Local Authority areas in London and this is the result:

![map_param_KMEANS](https://user-images.githubusercontent.com/7291547/156884047-33587e37-33e1-453a-bb55-2634753e788e.png)

Here's the code:

```
library(rmap)
library(tidyverse)
library(readxl)
library(janitor)

# read in new test data
download.file("https://data.london.gov.uk/download/gcse-results-by-borough/a8a71d73-cc48-4b30-9eb5-c5f605bc845c/gcse-results.xlsx",
              destfile = "gcse-results.xlsx")
gcse_results <- read_xlsx("gcse-results.xlsx", sheet = "2020-21")

# clean up test data
colnames <- paste0(gcse_results[1,], gcse_results[2,])
colnames <- gsub("NA", "", colnames)
names(gcse_results) <- colnames
gcse_results <- gcse_results %>%
  clean_names() %>%
  slice(4:36) %>%
  mutate(number_of_pupils_at_the_end_of_key_stage_4 = as.numeric(number_of_pupils_at_the_end_of_key_stage_4))

# try to map using Local Authority name
my_map <- rmap::map(gcse_results,
          subRegCol = "area",
          valueCol = "number_of_pupils_at_the_end_of_key_stage_4")

```

Again I know you can manually find a better fitting map, but then once again I'm not certain the contribution over finding a map manually in other ropensci packages like rnaturalearth and using tmap:

```
library(rnaturalearth)
library(tmap)

london_boroughs <- ne_states(country = "United Kingdom", returnclass = "sf") %>%
  filter(region == "Greater London")

gcse_results_sf <- left_join(london_boroughs, gcse_results,
                             by = c("name" = "area"))

qtm(gcse_results_sf, fill = "number_of_pupils_at_the_end_of_key_stage_4")

```
which returns:

![tmap_london](https://user-images.githubusercontent.com/7291547/156884587-2a52745d-88f9-4f3c-8566-7bc0a100ce15.png)

Or if I want the legend fixed:

```
qtm(gcse_results_sf, fill = "number_of_pupils_at_the_end_of_key_stage_4",  fill.title="Pupils at KS4")
```

![tmap_w_title](https://user-images.githubusercontent.com/7291547/156884596-607661ff-bc07-4f7b-a952-72f098911020.png)

In sum: I think that there is some more extensitve and thorough testing required than something that should be done by reviewers. I'm just spot checking here and keep coming across issues - so at this point I would advise some more thorough testing of the package overall by the authors, before I would be happy to sign off on this.
