### 2022-03-09 7a89daa
Updating state map and DiffAbs
