### 2022-03-17 df12f82
Switching rmap to sf package
