+++ inst/extras/saveDataFilesDF.R
@@ -1,478 +1,0 @@
-#' saveDataFiles.R
-
-#---------------------------
-# Libraries Needed
-#---------------------------
-library(data.table)
-library(usethis)
-library(dplyr)
-library(rmap)
-
-
-#-------------------
-# Save Maps as ggplot dataframes
-#-------------------
-
-# Worldmap countries
-#-------------------
-mapx <- rmap::mapCountries
-mapCountriesdf <- rmap::tidy_shape(shape=mapx, shapeColum="subRegion") %>%
-  dplyr::rename(subRegion=id)%>%
-  dplyr::inner_join(mapx@data, by="subRegion") %>% dplyr::rename(lon=long) %>% dplyr::mutate(name = paste0(name,"df"));
-mapCountriesdf %>% head()
-use_data(mapCountriesdf, version=3,  overwrite=T)
-
-# Worldmap states
-#-------------------
-mapx <- rmap::mapStates
-mapStatesdf <- rmap::tidy_shape(shape=mapx, shapeColum="subRegion") %>%
-  dplyr::rename(subRegion=id)%>%
-  dplyr::inner_join(mapx@data, by="subRegion") %>% dplyr::rename(lon=long) %>% dplyr::mutate(name = paste0(name,"df"));
-mapStatesdf %>% head()
-use_data(mapStatesdf, version=3,  overwrite=T)
-
-#-----------------
-# GCAM Maps (Regions, Basins, Land)
-#-----------------
-
-# GCAM 32 Regions
-#------------------
-mapx <- rmap::mapGCAMReg32
-mapGCAMReg32df <- rmap::tidy_shape(shape=mapx, shapeColum="subRegion") %>%
-  dplyr::rename(subRegion=id)%>%
-  dplyr::inner_join(mapx@data, by="subRegion") %>% dplyr::rename(lon=long) %>% dplyr::mutate(name = paste0(name,"df"));
-mapGCAMReg32df %>% head()
-use_data(mapGCAMReg32df, version=3,  overwrite=T)
-
-# GCAM Basins
-#------------------
-mapx <- rmap::mapGCAMBasins
-mapGCAMBasinsdf <- rmap::tidy_shape(shape=mapx, shapeColum="subRegion") %>%
-  dplyr::rename(subRegion=id)%>%
-  dplyr::inner_join(mapx@data, by="subRegion") %>% dplyr::rename(lon=long) %>% dplyr::mutate(name = paste0(name,"df"));
-mapGCAMBasinsdf %>% head()
-use_data(mapGCAMBasinsdf, version=3,  overwrite=T)
-
-# GCAM Land
-#------------------
-mapx <- rmap::mapGCAMLand
-mapGCAMLanddf <- rmap::tidy_shape(shape=mapx, shapeColum="subRegion") %>%
-  dplyr::rename(subRegion=id)%>%
-  dplyr::inner_join(mapx@data, by="subRegion") %>% dplyr::rename(lon=long) %>% dplyr::mutate(name = paste0(name,"df"));
-mapGCAMLanddf %>% head()
-use_data(mapGCAMLanddf, version=3,  overwrite=T)
-
-#-----------------
-# Hydrology Maps (HydroShed, HUC)
-#-----------------
-# Hydro sheds
-# https://www.hydrosheds.org/page/hydrobasins
-# Lehner, B., Grill G. (2013): Global river hydrography and network routing:
-# baseline data and new approaches to study the world’s large river systems.
-# Hydrological Processes, 27(15): 2171–2186. Data is available at www.hydrosheds.org
-
-# HydroSheds Level 1
[…]
