+++ R/map_plot.R
@@ -4,14 +4,14 @@
 #' Each figure is accompanied with a csv table.
 #'
 #' @param data Default = NULL
+#' @param region Default = NULL. Set the boundary region for subRegion maps. Useful when multiple subRegions in different regions.
 #' @param fileName Default = "map"
 #' @param shape Default = NULL, Cusotm shape can be provided as a SpatialPolygonDataFrame with features corresponding to subRegion columns in the data provided.
 #' @param folder Default = paste(getwd(),"/outputs",sep Default = "")
-#' @param palette Default = "Spectral"
+#' @param palette Default = "Set3"
 #' @param show Default = T. Print maps in console as they are processed.
 #' @param theme Default = NULL
 #' @param fillColumn Default = NULL # Or give column with data
-#' @param shapeColumn (Optional). Default = NULL. If different from subRegion.
 #' @param width Default = 9
 #' @param height Default = 7
 #' @param legendShow Default = F
@@ -20,6 +20,7 @@
 #' @param legendBreaksn Default = "5"
 #' @param legendBreaks Default = NULL
 #' @param labels Default = FALSE
+#' @param labelCol Default = NULL,
 #' @param labelRepel Default = 0,
 #' @param labelColor Default = "black",
 #' @param labelSize Default = 3
@@ -30,11 +31,13 @@
 #' @param color Default = "grey40". Color of polygon lines.
 #' @param lwd Default = 0.1. Line width of polygon boundaries.
 #' @param underLayer Default = NULL
+#' @param underLayerLabelCol Default = NULL
 #' @param underLayerColor Default = "gray40"
 #' @param underLayerFill Default = "gray40"
 #' @param underLayerLwd Default = 0.5
 #' @param underLayerAlpha Default = 1
 #' @param underLayerLabels Default = F
+#' @param overLayerLabelCol Default = NULL
 #' @param overLayerLabels Default = F
 #' @param overLayer Default = NULL
 #' @param overLayerColor Default = "gray40"
@@ -67,22 +70,25 @@
 #' @param crop_to_overLayer Default = F. Crop to the overLayer boundary provided.
 #' @param transparent Default = F. To make map background transparent for maps without backgrounds.
 #' @param legendType Default = "continuous".
+#' @param crs Default = 4326, WGS84
 #' @keywords charts, diffplots
 #' @return Returns the formatted data used to produce chart
+#' @import sf
 #' @importFrom rlang :=
 #' @importFrom jgcricolors jgcricol
 #' @importFrom magrittr %>%
 #' @export
 
 map_plot<-function(data=NULL,
+                  region=NULL,
                   fillColumn=NULL, # Or give column data with
-                  shapeColumn=NULL,
                   shape = NULL,
                   theme = NULL,
                   show = T,
-                  palette="Spectral",
+                  palette="Set3",
                   legendType="kmeans",
                   labels=F,
+                  labelCol = NULL,
                   labelRepel = 0,
                   labelColor = "black",
                   labelSize = 2,
@@ -99,18 +105,20 @@ map_plot<-function(data=NULL,
                   underLayer = NULL,
                   color = "grey40",
                   lwd = 0.1,
+                  underLayerLabelCol = NULL,
                   underLayerColor = "gray40",
[…]
