+++ vignettes/vignette_map.Rmd
@@ -778,7 +778,7 @@ rmap::map(data,
 
 For example if someone wants to analyze counties in Texas.
 
-```{r, results = 'hide', eval=TRUE, echo=TRUE, warning=FALSE, error = FALSE, message = FALSE}
+```{r, results = 'hide', eval=F, echo=TRUE, warning=FALSE, error = FALSE, message = FALSE}
 library(rmap); library(raster); library(sf)
 
 shapeSubRegions <- rmap::mapUS49County
