+++ NAMESPACE
@@ -4,7 +4,6 @@ export(map)
 export(map_find)
 export(map_plot)
 export(printPdfPng)
-import(sf)
 importFrom(jgcricolors,jgcricol)
 importFrom(magrittr,"%>%")
 importFrom(rlang,":=")
