+++ R/map.R
@@ -412,7 +412,7 @@ map <- function(data = NULL,
       underLayerLwd = underLayerLwd,
       underLayerAlpha = underLayerAlpha,
       background=background,
-      zoom=zoom, zoomx = zoomx, zoomy=zoomy, asp=asp, legendShow=legendShow,  crop = crop, crop_to_underLayer = crop_to_underLayer, crop_to_overLayer = crop_to_overLayer, transparent=transparent,
+      zoom=zoom, zoomx = zoomx, zoomy=zoomy, asp=asp, legendShow=F,  crop = crop, crop_to_underLayer = crop_to_underLayer, crop_to_overLayer = crop_to_overLayer, transparent=transparent,
       alpha = alpha, size=max(1,(size+(colm+rowm)*3 - 12)),
       theme = theme,
       legendTitle=NULL,
