+++ NAMESPACE
@@ -1,9 +1,6 @@
 # Generated by roxygen2: do not edit by hand
 
-export(map)
-export(map_find)
-export(map_plot)
-export(printPdfPng)
+import(sf)
 importFrom(jgcricolors,jgcricol)
 importFrom(magrittr,"%>%")
 importFrom(rlang,":=")
