+++ .github/workflows/build.yml
@@ -14,7 +14,7 @@ jobs:
       fail-fast: false
       matrix:
         config:
-        - { os: windows-latest, r: 'devel'}
+        - { os: ubuntu-latest, r: 'devel'}
     
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true      
