+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.2
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2022-03-09T19:16Z
+last_built: 2022-03-18T11:18Z
 
