### 2022-03-18 0c2b8e5
Updating incorrect maps and crop
