### 2022-03-19 987e1db
Updating tests to increase coverage of map_plot
