+++ tests/testthat/test-map.R
@@ -1,5 +1,5 @@
 context("map Tests")
-library(rmap); library(testthat); library(dplyr); library(sf)
+library(rmap); library(testthat); library(dplyr); library(sf); library(raster)
 
 test_that("map plots shape correctly", {
 
@@ -182,5 +182,201 @@ test_that("test multi-row multi-col", {
 })
 
 
+test_that("test input sf data with shape specified", {
+  mapx <- rmap::map(data=rmap::mapUS49 %>% dplyr::mutate(value=1:n()),
+                    shape=rmap::mapUS52Compact,
+                    show =F, save = F)
+  tVal1 <- length(mapx)
+  testthat::expect_gt(tVal1,0)
+})
+
+test_that("test raster data", {
+  datax <- raster::rasterFromXYZ(rmap::example_gridData_GWPv4To2015%>%head(50)%>%dplyr::filter(x==1990)%>%dplyr::select(-x))
+  mapx <- rmap::map_plot(data=datax,
+                    show =F, save = F)
+  tVal1 <- length(mapx)
+  testthat::expect_gt(tVal1,0)
+})
+
+test_that("legend fixed breaks", {
+  data = data.frame(subRegion = c("Austria","Spain", "Italy", "Germany","Greece"),
+                    value = c(32, 38, 54, 63, 24))
+
+  mapx <- rmap::map(data = data, show=F,save=F,
+            legendFixedBreaks=c(30,32,1000))
+  tVal1 <- length(mapx)
+  testthat::expect_gt(tVal1,0)
+})
+
+test_that("legend continuous", {
+  data = data.frame(subRegion = c("Austria","Spain", "Italy", "Germany","Greece"),
+                    value = c(32, 38, 54, 63, 24))
+
+  mapx <- rmap::map(data = data, show=F,save=F,
+                    legendType = "continuous")
+  tVal1 <- length(mapx)
+  testthat::expect_gt(tVal1,0)
+})
+
+test_that("legend breaks number", {
+  data = data.frame(subRegion = c("Austria","Spain", "Italy", "Germany","Greece"),
+                    value = c(32, 38, 54, 63, 24))
+
+  mapx <- rmap::map(data = data, show=F,save=F,
+                    legendBreaksn = 3)
+  tVal1 <- length(mapx)
+  testthat::expect_gt(tVal1,0)
+})
+
+test_that("legend breaks number", {
+  data = data.frame(subRegion = c("Austria","Spain", "Italy", "Germany","Greece"),
+                    value = c(32, 38, 54, 63, 24))
+
+  mapx <- rmap::map(data = data, show=F,save=F,
+                    legendSingleValue = 54,
+                    legendSingleColor = "green")
+  tVal1 <- length(mapx)
+  testthat::expect_gt(tVal1,0)
+})
 
 
+test_that("showNA", {
+  data = data.frame(subRegion = c("Austria","Spain", "Italy", "Germany","Greece"),
+                    value = c(32, 38, 54, 63, NA))
+
+  mapx <- rmap::map(data = data, show=F,save=F,
+                    showNA=T)
+  tVal1 <- length(mapx)
+  testthat::expect_gt(tVal1,0)
+})
+
+test_that("region and underLayer gg", {
+  data = data.frame(subRegion = c("Punjab","Punjab"),
+                    region = c("India","Pakistan"),
+                    value = c(32, 38))
+
+  mapx <- rmap::map(data = data, show=F,save=F,
+                    underLayer=rmap::mapCountries%>%dplyr::filter(subRegion %in% c("India","China","Pakistan")))
+
+  mapx1 <- rmap::map(data = data, show=F,save=F,
+                    region="Pakistan",
[…]
