### 2022-03-19 db434f9
Updating tests to increase coverage of map_plot
