### 2022-03-20 6a163ea
Udpating projections with cropping
