+++ DESCRIPTION
@@ -36,7 +36,8 @@ Imports:
 	rgeos,
 	rgdal,
 	sf,
-	methods
+	methods,
+	ggrepel
 Suggests: 
     rmarkdown,
     testthat (>= 3.0.0),
