### 2022-03-20 59e249a
Udpating tests for map_plot
