+++ tests/testthat/test-map.R
@@ -89,16 +89,21 @@ test_that("map saves both pdf and png", {
 test_that("numeric2cat works", {
   # Create a list of ranges and categorical color scales for each parameter
   numeric2Cat_param <- list("param1",
-                            "param2")
+                            "param2",
+                            "param3")
   numeric2Cat_breaks <- list(c(-Inf, 0.1,1.1,2.1,3.1,4.1,5.1,10.1,Inf),
-                             c(-Inf, 0.1, 0.2, 0.4,Inf))
+                             c(-Inf, 0.1, 0.2, 0.4,Inf),
+                             c(-Inf, 0.1,1.1,2.1,3.1,4.1,5.1,10.1,Inf))
   numeric2Cat_labels <- list(c("0","1","2","3","4","5","10",">10"),
-                             c(names(jgcricolors::jgcricol()$pal_scarcityCat)))
+                             c(names(jgcricolors::jgcricol()$pal_scarcityCat)),
+                             c("0","1","2","3","4","5","10",">10"))
   numeric2Cat_palette <- list(c("0"="green","1"="#fee5d9","2"="#fcbba1",
                                 "3"="#fc9272","4"="#fb6a4a","5"="#de2d26",
                                 "10"="#a50f15",">10"="black"),
-                              c("pal_scarcityCat")) # Can be a custom scale or an R brewer palette or an rmap palette
+                              c("pal_scarcityCat"),
+                              c("0"="green")) # Can be a custom scale or an R brewer palette or an rmap palette
   numeric2Cat_legendTextSize <- list(c(0.7),
+                                     c(0.7),
                                      c(0.7))
   numeric2Cat_list <-list(numeric2Cat_param = numeric2Cat_param,
                           numeric2Cat_breaks = numeric2Cat_breaks,
@@ -107,18 +112,21 @@ test_that("numeric2cat works", {
                           numeric2Cat_legendTextSize = numeric2Cat_legendTextSize); numeric2Cat_list
 
   data = data.frame(subRegion = c("CA","AZ","TX","NH","ID","OH",
+                                  "CA","AZ","TX","NH","ID","OH",
                                   "CA","AZ","TX","NH","ID","OH"),
                     x = c(2050,2050,2050,2050,2050,2050,
+                          2050,2050,2050,2050,2050,2050,
                           2050,2050,2050,2050,2050,2050),
                     value = c(0,1,3,20,2,1,
-                              0,0.1,0.3,0.2,0.25,0.5),
-                    param = c(rep("param1",6),rep("param2",6)))
+                              0,0.1,0.3,0.2,0.25,0.5,
+                              0,1,3,20,2,1),
+                    param = c(rep("param1",6),rep("param2",6),rep("param3",6)))
 
   mapx <- rmap::map(data = data,
             background = T,
             underLayer = rmap::mapCountries,
             numeric2Cat_list = numeric2Cat_list,
-            show=F, save=F)
+            show=T, save=T)
   tVal1 <- length(mapx)
   testthat::expect_gt(tVal1,0)
 })
@@ -172,13 +180,18 @@ test_that("test multi-row multi-col", {
                               37, 53, 23, 12, 45,
                               23, 99, 102, 85, 75,
                               12, 76, 150, 64, 90))
-  mapx <- rmap::map(data = data, show=F, save=F,
-            underLayer = rmap::mapCountries,
-            row = "rcp",
-            col = "gcm",
-            background = T )
-  tVal1 <- length(mapx)
-  testthat::expect_gt(tVal1,0)
[…]
