+++ tests/testthat/test-map.R
@@ -163,171 +163,171 @@ test_that("test layers and labels", {
   testthat::expect_gt(tVal1,0)
 })
 
-test_that("test multi-row multi-col", {
-  data = data.frame(subRegion = c("Austria","Spain", "Italy", "Germany","Greece",
-                                  "Austria","Spain", "Italy", "Germany","Greece",
-                                  "Austria","Spain", "Italy", "Germany","Greece",
-                                  "Austria","Spain", "Italy", "Germany","Greece"),
-                    rcp = c(rep("RCP1",5),
-                            rep("RCP2",5),
-                            rep("RCP1",5),
-                            rep("RCP2",5)),
-                    gcm = c(rep("GCM1",5),
-                            rep("GCM1",5),
-                            rep("GCM2",5),
-                            rep("GCM2",5)),
-                    value = c(32, 38, 54, 63, 24,
-                              37, 53, 23, 12, 45,
-                              23, 99, 102, 85, 75,
-                              12, 76, 150, 64, 90))
-  mapx1 <- rmap::map(data = data, show=F, save=F,row = "rcp",col = "gcm")
-  mapx2 <- rmap::map(data = data, show=F, save=F,row = c("rcp","gcm"),col = c("rcp","gcm"))
-  mapx3 <- rmap::map(data = data, show=F, save=F,row = c("rcp","rcp","gcm"),col = c("rcp","rcp","gcm"))
-  mapx4 <- rmap::map(data = data%>%dplyr::filter(rcp=="RCP1"), show=F, save=F,row = "gcm")
-  mapx5 <- rmap::map(data = data, show=F, save=F,row = c("rcp","gcm"))
-  mapx6 <- rmap::map(data = data, show=F, save=F,row = c("rcp","rcp","gcm"))
-  mapx7 <- rmap::map(data = data%>%dplyr::filter(rcp=="RCP1"), show=F, save=F,col = "gcm")
-  mapx8 <- rmap::map(data = data, show=F, save=F,col = c("rcp","gcm"))
-  mapx9 <- rmap::map(data = data, show=F, save=F,col = c("rcp","rcp","gcm"))
-  testthat::expect_equal(sum(length(mapx1),length(mapx2),length(mapx3),
-                          length(mapx4),length(mapx5),length(mapx6),
-                          length(mapx7),length(mapx8),length(mapx9)),9)
-})
-
-
-test_that("test input sf data with shape specified", {
-  mapx <- rmap::map(data=rmap::mapUS49 %>% dplyr::mutate(value=1:n()),
-                    shape=rmap::mapUS52Compact,
-                    show =F, save = F)
-  tVal1 <- length(mapx)
-  testthat::expect_gt(tVal1,0)
-})
-
-test_that("test raster data", {
-  datax <- raster::rasterFromXYZ(rmap::example_gridData_GWPv4To2015%>%head(50)%>%dplyr::filter(x==1990)%>%dplyr::select(-x))
-  mapx <- rmap::map_plot(data=datax,
-                    show =F, save = F)
-  tVal1 <- length(mapx)
-  testthat::expect_gt(tVal1,0)
-})
-
-test_that("legend fixed breaks", {
-  data = data.frame(subRegion = c("Austria","Spain", "Italy", "Germany","Greece"),
-                    value = c(32, 38, 54, 63, 24))
-
-  mapx <- rmap::map(data = data, show=F,save=F,
-            legendFixedBreaks=c(30,32,1000))
-  tVal1 <- length(mapx)
-  testthat::expect_gt(tVal1,0)
-})
-
-test_that("legend continuous", {
[…]
