### 2022-03-21 #4015 @zarrarkhan
@maczokni @hugoledoux

Thank you for your earlier comments. We have tried to address all of your comments and made the following changes. Please, pull in the latest changes from `main` to test these and let us know if you have any additional comments.

- We have removed the dependency on the `sp` package and switched to `sf`. All `rmap` internal data maps are now `sf` objects
- We added the ability to change map projections: https://jgcri.github.io/rmap/articles/vignette_map.html#projections
- We improved test coverage of all underlying functions. Three of the total four `rmap` functions are now at >85% CI test coverage with over all package test coverage improved from 55% to 65%. (map.R 60.84%, map_plot.R 86%, map_find.R 92%, printPdfPng.R 91%)
- We tested the package for the example cases provided by @maczokni, our original example cases in our documentation, as well as additional test cases not in the documentation. Some of these are provided below:

**Load Libraries**
```
library(rmap); library(geodaData); library(dplyr); library(sf); library(tidyverse)
library(readxl); library(janitor)
```

**Example 1**
```
data("ncovr")
mydata <- ncovr %>%
  dplyr::select(NAME, STATE_NAME, FIPS, HR60) %>%
  sf::st_drop_geometry() %>%
  dplyr::left_join(rmap::mapUS52County, by="FIPS")
rmap::map(data=mydata, valueCol = "HR60", legendTitle = "HR60")
```
![map_param_KMEANS](https://user-images.githubusercontent.com/7863115/159287352-bfbf772a-5423-4be0-9456-22b3344686e9.png)

**Example 2**
```
download.file("https://data.london.gov.uk/download/gcse-results-by-borough/a8a71d73-cc48-4b30-9eb5-c5f605bc845c/gcse-results.xlsx", destfile = "gcse-results.xlsx")
gcse_results <- readxl::read_xlsx("gcse-results.xlsx", sheet = "2020-21")

# clean up test data
colnames <- paste0(gcse_results[1,], gcse_results[2,])
colnames <- gsub("NA", "", colnames)
names(gcse_results) <- colnames
gcse_results <- gcse_results %>%
  clean_names() %>%
  slice(4:36) %>%
  mutate(number_of_pupils_at_the_end_of_key_stage_4 = as.numeric(number_of_pupils_at_the_end_of_key_stage_4))

# Plot
my_map <- rmap::map(gcse_results,
                    subRegCol = "area",
                    valueCol = "number_of_pupils_at_the_end_of_key_stage_4",
                    legendTitle = "Pupils at KS4")
```
![map_param_KMEANS](https://user-images.githubusercontent.com/7863115/159287912-c7f5dc76-c481-43f5-9996-8d04314b9b8a.png)

**Example 3**
```
# Test Covid data
# Our World in Data JHU https://github.com/owid/covid-19-data/tree/master/public/data
# State vaccination data: https://github.com/owid/covid-19-data/raw/master/public/data/vaccinations/us_state_vaccinations.csv
covid_data <- read.csv(url("https://github.com/owid/covid-19-data/raw/master/public/data/vaccinations/us_state_vaccinations.csv"))%>%
  tibble::as_tibble() %>%
  dplyr::select(subRegion=location,date,value=people_vaccinated_per_hundred) %>%
  dplyr::mutate(subRegion = dplyr::if_else(subRegion=="New York State","New York",subRegion)) %>%
  dplyr::filter(date == max(date),
                subRegion %in% rmap::mapUS49$subRegionAlt); covid_data

rmap::map(covid_data,
          title=paste0("People Vaccinated per hundered ",max(covid_data$date)),
          legendTitle = "People per 100")
```
![map_param_KMEANS](https://user-images.githubusercontent.com/7863115/159287656-588a024f-c928-42b6-a3d6-ba7028dc549a.png)

**Examples from Documentation**: https://jgcri.github.io/rmap/articles/vignette_map.html
