### 2022-04-05 #4015 @zarrarkhan
Hi @hugoledoux,

Just wanted to check in to see if there were any updates since we submitted our modifications in https://github.com/openjournals/joss-reviews/issues/4015#issuecomment-1074078464 two weeks ago.

Thanks!
