### 2022-04-15 cbf54a1
Fixing issue with EU-12 and EU-15 in response to https://github.com/JGCRI/rmap/issues/90#issue-1204582283
