### 2022-04-15 c65d6f6
Upating Uruguay maps to high res for https://github.com/JGCRI/rmap/issues/89#issue-1204576875
