+++ paper/paper.bib
@@ -13,8 +13,7 @@
 	author = {Wild, Thomas B. and Khan, Zarrar and Clarke, Leon and Hejazi, Mohamad and Bereslawski, Julia Lacal and Suriano, Micaela and Roberts, Paula and Casado, José and Miralles-Wilhelm, Fernando and Gavino-Novillo, Marcelo and Muñoz-Castillo, Raul and Moreda, Fekadu and Zhao, Mengqi and Yarlagadda, Brinda and Lamontagne, Jonathan and Birnbaum, Abigail},
 	month = jun,
 	year = {2021},
-	pages = {62},
-	file = {Springer Full Text PDF:C\:\\Users\\khan404\\Zotero\\storage\\PCUAYYCW\\Wild et al. - 2021 - Integrated energy-water-land nexus planning in the.pdf:application/pdf},
+	pages = {62}
 }
 
 @article{khan_future_2021,
@@ -32,8 +31,7 @@
 	month = jun,
 	year = {2021},
 	note = {Publisher: IOP Publishing},
-	pages = {065010},
-	file = {IOP Full Text PDF:C\:\\Users\\khan404\\Zotero\\storage\\B9WHSREE\\Khan et al. - 2021 - The future evolution of energy-water-agriculture i.pdf:application/pdf},
+	pages = {065010}
 }
 
 @article{tennekes_tmap_2018,
@@ -59,8 +57,7 @@
 	journal = {Journal of Open Source Software},
 	author = {Giraud, Timothée and Lambert, Nicolas},
 	year = {2016},
-	pages = {54},
-	file = {Full Text:C\:\\Users\\khan404\\Zotero\\storage\\BW9BV6FE\\openurl.pdf:application/pdf},
+	pages = {54}
 }
 
 @article{south_rworldmap_2011,
@@ -72,16 +69,14 @@
 	number = {1},
 	journal = {R Journal},
 	author = {South, Andy},
-	year = {2011},
-	file = {Full Text:C\:\\Users\\khan404\\Zotero\\storage\\UWVEBC8U\\South - 2011 - rworldmap a new R package for mapping global data.pdf:application/pdf},
+	year = {2011}
 }
 
 @article{brunsdon_package_2015,
 	title = {Package ‘{GISTools}’},
 	journal = {CRAN Repository},
-	author = {Brunsdon, Chris and Chen, Hongyan and Brunsdon, Maintainer Chris},
-	year = {2015},
-	file = {Full Text:C\:\\Users\\khan404\\Zotero\\storage\\B9Q8QVLF\\Brunsdon et al. - 2015 - Package ‘GISTools’.pdf:application/pdf},
+	author = {Brunsdon, Chris and Chen, Hongyan},
+	year = {2015}
 }
 
 @misc{qgis_development_team_qgis_2021,
@@ -98,8 +93,7 @@
 	shorttitle = {Open source {GIS}},
 	publisher = {Springer Science \& Business Media},
 	author = {Neteler, Markus and Mitasova, Helena},
-	year = {2013},
-	file = {Full Text:C\:\\Users\\khan404\\Zotero\\storage\\TPBIZTF7\\Neteler and Mitasova - 2013 - Open source GIS a GRASS GIS approach.pdf:application/pdf;Snapshot:C\:\\Users\\khan404\\Zotero\\storage\\9R9EK3BH\\books.html:text/html},
+	year = {2013}
 }
 
 @article{pebesma_s_2005,
@@ -111,8 +105,7 @@
 	journal = {R news},
 	author = {Pebesma, Edzer and Bivand, Roger S.},
 	year = {2005},
-	pages = {9--13},
[…]
