### 2022-04-22 #4015 @zarrarkhan
@hugoledoux Thank you for generating the proof. We have read the proof and corrected a few typos we found.

In addition we see that the referencing style seems inconsistent but it is not clear how to adjust this. For example: In the article proof linked above on line 22 we have the following:

```
publications (Khan et al., 2021; Wild, Khan, Clarke, et al., 2021; Wild, Khan, Zhao, et al., 2021)
```
Which refer to the following references:
- Line 118: Khan, Z., Wild, T. B., Iyer, G., Hejazi, M., & Vernon, C. R. (2021). The future evolution of energy-water-agriculture interconnectivity across the US. Environmental Research Letters,16(6), 065010. https://iopscience.iop.org/article/10.1088/1748-9326/ac046c
- Line 140: Wild, T. B., Khan, Z., Clarke, L., Hejazi, M., Bereslawski, J. L., Suriano, M., Roberts, P., Casado, J., Miralles-Wilhelm, F., Gavino-Novillo, M., Muñoz-Castillo, R., Moreda, F., Zhao, M., Yarlagadda, B., Lamontagne, J., & Birnbaum, A. (2021). Integrated energy-water-land nexus planning in the Colorado River Basin (Argentina). Regional Environmental Change, 21(3), 62. https://doi.org/10.1007/s10113-021-01775-1
- Line 145: Wild, T. B., Khan, Z., Zhao, M., Suriano, M., Bereslawski, J. L., Roberts, P., Casado, J., Gaviño-Novillo, M., Clarke, L., & Hejazi, M. (2021). The Implications of Global Change for the Co-Evolution of Argentina’s Integrated Energy-Water-Land Systems. Earth’s Future, 9(8), e2020EF001970. https://doi.org/10.1029/2020EF001970

The first citation only references a single author while the rest list three authors.

Apart from this we have checked the proof and pushed all changes in commit: https://github.com/JGCRI/rmap/commit/6730a68880e6673584d67b45e66820ee1fb86d9a

Please let us know what we can do next.

Thanks!
