+++ inst/extras/saveDataFiles.R
@@ -39,6 +39,7 @@ if(redoMaps){
     dplyr::mutate(name="mapCountries") %>%
     dplyr::select(subRegion,region,subRegionAlt,source, geometry, area, name,
                   pop_est, gdp_md_est, pop_year, lastcensus, gdp_year)
+  sf::st_crs(mapCountries) <- 4326
   use_data(mapCountries, version=3, overwrite=T)
 }
 
@@ -82,7 +83,7 @@ if(redoMaps){
   mapStates <- mapStates %>%
     dplyr::mutate(name="mapStates") %>%
     dplyr::select(subRegion,region,subRegionAlt,source, area, name, geometry)
-
+  sf::st_crs(mapStates) <- 4326
   use_data(mapStates, version=3, overwrite=T)
 
   # Check
@@ -145,6 +146,7 @@ if(redoMaps){
     dplyr::mutate(name="mapGCAMReg32")%>%
     dplyr::select(subRegion,region,subRegionAlt,source, area, name, geometry)%>%
     st_transform(st_crs(4326))
+  sf::st_crs(mapGCAMReg32) <- 4326
   use_data(mapGCAMReg32, version=3, overwrite=T)
 
 }
@@ -192,6 +194,7 @@ if(redoMaps){
     dplyr::mutate(name="mapGCAMBasins")%>%
     dplyr::select(subRegion,region,subRegionAlt,source, area, name, geometry)%>%
     sf::st_transform(st_crs(4326))
+  sf::st_crs(mapGCAMBasins) <- 4326
   use_data(mapGCAMBasins, version=3, overwrite=T)
 }
 
@@ -249,6 +252,7 @@ if(redoMaps){
                   subRegion_GCAMBasin, subRegionAlt_GCAMBasin,
                   subRegion_GCAMReg32, subRegionAlt_GCAMReg32)%>%
     st_transform(st_crs(4326))
+  sf::st_crs(mapGCAMLand) <- 4326
   use_data(mapGCAMLand, version=3, overwrite=T)
 }
 
@@ -288,6 +292,7 @@ if(redoMaps){
   mapHydroShed1 <- mapHydroShed1 %>%
     dplyr::mutate(name="mapHydroShed1")%>%
     dplyr::select(subRegion,region,subRegionAlt,source, area, name, geometry)
+  sf::st_crs(mapHydroShed1) <- 4326
   use_data(mapHydroShed1, version=3, overwrite=T)
 }
 
@@ -314,6 +319,7 @@ if(redoMaps){
   mapHydroShed2 <- mapHydroShed2 %>%
     dplyr::mutate(name="mapHydroShed2")%>%
     dplyr::select(subRegion,region,subRegionAlt,source, area, name, geometry)
+  sf::st_crs(mapHydroShed2) <- 4326
   use_data(mapHydroShed2, version=3, overwrite=T)
 }
 
@@ -341,6 +347,7 @@ if(redoMaps){
   mapHydroShed3 <- mapHydroShed3 %>%
     dplyr::mutate(name="mapHydroShed3")%>%
     dplyr::select(subRegion,region,subRegionAlt,source, area, name, geometry)
+  sf::st_crs(mapHydroShed3) <- 4326
   use_data(mapHydroShed3, version=3, overwrite=T)
 }
 
@@ -382,6 +389,7 @@ if(redoMaps){
     dplyr::mutate(name="mapUS52HUC2")%>%
     dplyr::select(subRegion,region,subRegionAlt,source, area, name, geometry) %>%
[…]
