+++ pkgdown.yml
@@ -1,7 +1,7 @@
 pandoc: 2.7.3
-pkgdown: 2.0.2
+pkgdown: 2.0.3
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2022-04-22T20:25Z
+last_built: 2022-05-05T20:08Z
 
