+++ DESCRIPTION
@@ -35,7 +35,7 @@ Imports:
 	cowplot,
 	rgeos,
 	rgdal,
-	sf,
+	sf (>= 1.0-7),
 	methods,
 	ggrepel
 Suggests: 
