+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.3
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2022-05-05T20:08Z
+last_built: 2022-05-19T00:08Z
 
