+++ R/map_plot.R
@@ -352,7 +352,6 @@ if(T){ # Read input data
   } else if(any(grepl("data.frame", class(data))) & any(grepl("^lat$",names(data))) & any(grepl("^lon$",names(data)))){
     # If simple dataframe with lat lon
     data_sf_raster <- raster::rasterFromXYZ(data %>%
-                                              dplyr::select("lon", "lat", dplyr::all_of(col), "value") %>%
                                               tidyr::spread(key=col,value="value"))
     data_sf_spdf <- methods::as(data_sf_raster,'SpatialPolygonsDataFrame')
     data_sf <- sf::st_as_sf(data_sf_spdf) %>%
