+++ DESCRIPTION
@@ -16,25 +16,25 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.1
 Imports: 
-	raster,
-	tibble,
-	tidyr,
-	dplyr, 
+	raster (>= 3.4-13),
+	tibble (>= 3.1.2),
+	tidyr (>= 1.1.3),
+	dplyr (>= 1.0.7), 
 	scales,
 	grDevices,
 	data.table,
 	graphics,
-	rlang, 
+	rlang (>= 0.4.11), 
 	magrittr,
 	RColorBrewer,
 	magick,
-	stringr,
-	ggplot2,
+	stringr (>= 1.4.0),
+	ggplot2 (>= 3.3.4),
     markdown,
 	jgcricolors,
-	cowplot,
-	rgeos,
-	rgdal,
+	cowplot (>= 1.1.1),
+	rgeos (>= 0.5-5),
+	rgdal (>= 1.5-23),
 	sf (>= 1.0-7),
 	methods,
 	ggrepel
