+++ DESCRIPTION
@@ -16,26 +16,26 @@ Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.1
 Imports: 
-	raster (>= 3.4-13),
-	tibble (>= 3.1.2),
-	tidyr (>= 1.1.3),
-	dplyr (>= 1.0.7), 
+	raster,
+	tibble,
+	tidyr,
+	dplyr, 
 	scales,
 	grDevices,
 	data.table,
 	graphics,
-	rlang (>= 0.4.11), 
+	rlang, 
 	magrittr,
 	RColorBrewer,
 	magick,
-	stringr (>= 1.4.0),
-	ggplot2 (>= 3.3.4),
+	stringr,
+	ggplot2,
     markdown,
 	jgcricolors,
-	cowplot (>= 1.1.1),
-	rgeos (>= 0.5-5),
-	rgdal (>= 1.5-23),
-	sf (>= 1.0-7),
+	cowplot,
+	rgeos,
+	rgdal,
+	sf,
 	methods,
 	ggrepel
 Suggests: 
