+++ .github/workflows/build.yml
@@ -14,7 +14,7 @@ jobs:
       fail-fast: false
       matrix:
         config:
-        - { os: ubuntu-latest, r: 'devel'}
+        - { os: windows-latest, r: 'devel'}
     
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true      
@@ -34,21 +34,6 @@ jobs:
           path: ${{ env.R_LIBS_USER }} 
           key: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-${{ hashFiles('.github/depends.Rds') }} 
           restore-keys: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1- 
-      
-      - name: Ubuntu config
-        if: startsWith(matrix.config.os, 'ubuntu')
-        run: |
-          sudo add-apt-repository ppa:ubuntugis/ppa
-          sudo apt-get update
-          sudo apt-get install -y libcurl4-openssl-dev libssl-dev libxml2-dev libudunits2-dev libgdal-dev libgeos-dev libproj-dev libavfilter-dev  libmagick++-dev
-      
-      - name: MacOS Config
-        if: startsWith(matrix.config.os, 'macOS')
-        run: |
-            brew install pkg-config
-            brew install gdal
-            brew install geos
-            brew install imagemagick@6
       
       - name: Install dependencies
         run: |
