### 2022-05-20 90b24c6
Updating scripts to work with linux systems
