+++ vignettes/vignette_map.Rmd
@@ -1093,23 +1093,23 @@ rmap::map(data = data,
 ### Gridded Data Multi-Class
 
 ```{r, results = 'hide', eval=T, echo=TRUE, warning=FALSE, error = FALSE, message = FALSE}
-# library(rmap);library(dplyr)
-# 
-# # Change the years in the gridded data to classes
-# data = example_gridData_GWPv4To2015 %>%
-#   dplyr::rename(class=x) %>%
-#   dplyr::mutate(class = dplyr::case_when(
-#     class == "2000" ~ "municipal",
-#     class == "2005" ~ "agriculture",
-#     class == "2010" ~ "water",
-#     class == "2015" ~ "other",
-#     TRUE ~ class
-#   ))
-#   
-# rmap::map(data = data,
-#           overLayer = rmap::mapCountries,
-#           background = T,
-#           ncol = 2, width = 10, height = 10)
+library(rmap);library(dplyr)
+
+# Change the years in the gridded data to classes
+data = example_gridData_GWPv4To2015 %>%
+  dplyr::rename(class=x) %>%
+  dplyr::mutate(class = dplyr::case_when(
+    class == "2000" ~ "municipal",
+    class == "2005" ~ "agriculture",
+    class == "2010" ~ "water",
+    class == "2015" ~ "other",
+    TRUE ~ class
+  ))
+
+rmap::map(data = data,
+          overLayer = rmap::mapCountries,
+          background = T,
+          ncol = 2, width = 10, height = 10)
 ```
 
 
@@ -1355,25 +1355,25 @@ mapx$map_param_KMEANS_DiffPrcnt
 library(rmap);
 
 # # Change the years in the example gridded data to scenarios
-# data = example_gridData_GWPv4To2015 %>%
-#   dplyr::filter(x %in% c("2000","2005","2010")) %>%
-#   dplyr::rename(scenario=x) %>%
-#   dplyr::mutate(scenario = dplyr::case_when(
-#     scenario == "2000" ~ "scenario1",
-#     scenario == "2005" ~ "scenario2",
-#     scenario == "2010" ~ "scenario3",
-#     TRUE ~ scenario))
-#   
-# mapx <- rmap::map(data = data, 
-#           underLayer = rmap::mapCountries,
-#           scenRef = "scenario1",
-#           scenDiff = c("scenario2","scenario3"),
-#           background = T,
-#           save=F, show=F,
-#           ncol=1)
-# 
-# mapx$map_param_KMEANS_DiffAbs
-# mapx$map_param_KMEANS_DiffPrcnt
+data = example_gridData_GWPv4To2015 %>%
+  dplyr::filter(x %in% c("2000","2005","2010")) %>%
+  dplyr::rename(scenario=x) %>%
+  dplyr::mutate(scenario = dplyr::case_when(
+    scenario == "2000" ~ "scenario1",
+    scenario == "2005" ~ "scenario2",
+    scenario == "2010" ~ "scenario3",
+    TRUE ~ scenario))
+
+mapx <- rmap::map(data = data,
+          underLayer = rmap::mapCountries,
+          scenRef = "scenario1",
+          scenDiff = c("scenario2","scenario3"),
+          background = T,
+          save=F, show=F,
+          ncol=1)
+
+mapx$map_param_KMEANS_DiffAbs
+mapx$map_param_KMEANS_DiffPrcnt
 
 ```
 
