### 2022-05-20 e0b5063
Updating scripts to work with linux systems
