### 2022-05-20 9afef4d
Updating scripts to work with linux systems
