+++ R/map_plot.R
@@ -548,199 +548,285 @@ if(T){ # Read input data
       }
 
     # Adding in a single value (eg. 0 to be set to a single color eg. white)
-      if(!is.null(legendDigits)){
-          if(length(legendBreaks)>1){
+    if (!is.null(legendDigits)) {
+      if (length(legendBreaks) > 1) {
 
-            # Test Palette
-            if(T){
-              # Legend Labels
-              if(T){
+        # Test Palette
+        if (T) {
 
-                legendBreaksX <- legendBreaks; legendBreaksX
+          # Legend Labels
+          if (T) {
+            legendBreaksX <- legendBreaks
+            legendBreaksX
 
-                # New Breaks
-                if(!is.numeric(legendSingleValue)){
-                  if(max(legendBreaksX)<0){
-                    legendSingleValueX=ceiling(max(legendBreaksX))
-                  }else{
-                    if(min(legendBreaksX)>0){
-                      legendSingleValueX=floor(min(legendBreaksX))
-                    }else{
-                      legendSingleValueX=0
-                    }
-                  }
-                }else {
-                  legendSingleValueX=legendSingleValue}; legendSingleValueX
-
-
-                # Place Single Value in Legend Breaks if not present
-                if(!legendSingleValueX %in% legendBreaksX){
-                  if(min(legendBreaksX)>legendSingleValueX){
-                    legendBreaksX <- c(legendSingleValueX,legendBreaksX)
-                  }else{
-                    if(max(legendBreaksX)<legendSingleValueX){
-                      legendBreaksX <- c(legendBreaksX,legendSingleValueX)
-                    }else{
-                      if(max(legendBreaksX)==legendSingleValueX | min(legendBreaksX)==legendSingleValueX){
-                        legendBreaksX <- legendBreaksX
-                      }else{
-                        legendBreaksX <- c(legendBreaksX[1:match(max(legendBreaksX[legendBreaksX<legendSingleValueX]),legendBreaksX)],
-                                           legendSingleValueX,
-                                           legendBreaksX[(match(max(legendBreaksX[legendBreaksX<legendSingleValueX]),legendBreaksX)+1):(length(legendBreaksX))])
-                      }
-                    }
-                  }
+            # New Breaks
+            if (!is.numeric(legendSingleValue)) {
+              if (max(legendBreaksX) < 0) {
+                legendSingleValueX = ceiling(max(legendBreaksX))
+              } else{
+                if (min(legendBreaksX) > 0) {
+                  legendSingleValueX = floor(min(legendBreaksX))
+                } else{
+                  legendSingleValueX = 0
                 }
-                singleValLoc <- max(match(legendSingleValueX,legendBreaksX),1); singleValLoc; legendBreaksX
-
-                # Legend Labels
-                a<-c()
-                if(length(legendBreaksX)>1){
-                  for(i in 1:(length(legendBreaksX)-1)){
-                    if(i!=1){lower<-upperLast}else{lower <- round(legendBreaksX[i],(legendDigits))};lower
-                    upper <- round(legendBreaksX[i+1],legendDigits); upper
-                    countDig <- 1
-                    while(upper==lower & countDig<6){upper <- round(legendBreaksX[i+1],(legendDigits+countDig)); countDig=countDig+1};upper
-                    upperLast <- upper; upperLast
-                    a[i]=paste(format(lower,big.mark = ",")," to ",format(upper,big.mark = ","),sep="")
-                  };a
-
-                  # Add in the single Value
-                  if(min(legendBreaksX)>=legendSingleValueX){
-                    legendLabelsX=c(a)
-                    legendLabelsX<-c(paste(legendSingleValueX,sep=""),
-                                     legendLabelsX)
-                  }else{
[…]
