### 2022-05-20 0f6d2bc
Updating scripts to fix continuous legend error
