+++ vignettes/vignette_map.Rmd
@@ -1040,7 +1040,7 @@ mapx <- rmap::map(data = data,
           background = T, 
           underLayer = rmap::mapCountries,
           xRef = 2000, 
-          xDiff = c(2010,2015))
+          xDiff = c(2010,2015), diffOnly = T)
 
 mapx$map_param_KMEANS_xDiffAbs
 mapx$map_param_KMEANS_xDiffPrcnt
