+++ .github/workflows/build.yml
@@ -23,8 +23,6 @@ jobs:
     steps:
       - uses: actions/checkout@v1
       - uses: r-lib/actions/setup-r@master
-        with:
-          r-version: '3.6'
       - uses: r-lib/actions/setup-pandoc@master
       - uses: r-lib/actions/setup-tinytex@master
       
