### 2022-05-20 db50f05
Updating to github actions for windows
