+++ .github/workflows/docs.yaml
@@ -8,7 +8,7 @@ on:
 
 jobs:
   pkgdown:
-    runs-on: ubuntu-latest
+    runs-on: windows-latest
     
     env:
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
