+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.3
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2022-05-19T01:12Z
+last_built: 2022-05-21T02:08Z
 
