### 2022-05-21 dca6f50
Removing X from grid map labels
