### 2022-05-21 b87c566
Removing X from grid map labels
