+++ .github/workflows/build.yml
@@ -5,24 +5,35 @@ on:
     branches:
       - main
 
-
 jobs:
-  check:
-    runs-on: ubuntu-latest
+  R-CMD:
+    runs-on: ${{ matrix.config.os }}
+    name: ${{ matrix.config.os }} (${{ matrix.config.r }})
+    
+    strategy:
+      fail-fast: false
+      matrix:
+        config:
+        - { os: ubuntu-latest, r: 'devel', cran: "https://demo.rstudiopm.com/all/__linux__/xenial/latest"}
     
     env:
-      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
+      R_REMOTES_NO_ERRORS_FROM_WARNINGS: true      
       GITHUB_PAT: ${{ secrets.GH_PAT}}
-          
+
     steps:
-    
-      - uses: actions/checkout@v2
-
-      - uses: r-lib/actions/setup-pandoc@v1
-
-      - uses: r-lib/actions/setup-r@v1
+      - uses: actions/checkout@v1
+      - uses: r-lib/actions/setup-r@master
         with:
-          use-public-rspm: true
+          r-version: '3.6'
+      - uses: r-lib/actions/setup-pandoc@master
+      - uses: r-lib/actions/setup-tinytex@master
+      
+      - name: Cache R packages 
+        uses: actions/cache@v1 
+        with: 
+          path: ${{ env.R_LIBS_USER }} 
+          key: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-${{ hashFiles('.github/depends.Rds') }} 
+          restore-keys: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1- 
           
       - name: Ubuntu config
         if: startsWith(matrix.config.os, 'ubuntu')
