+++ .github/workflows/build.yml
@@ -27,13 +27,6 @@ jobs:
           r-version: '3.6'
       - uses: r-lib/actions/setup-pandoc@master
       - uses: r-lib/actions/setup-tinytex@master
-      
-      - name: Cache R packages 
-        uses: actions/cache@v1 
-        with: 
-          path: ${{ env.R_LIBS_USER }} 
-          key: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1-${{ hashFiles('.github/depends.Rds') }} 
-          restore-keys: ${{ runner.os }}-${{ hashFiles('.github/R-version') }}-1- 
           
       - name: Ubuntu config
         if: startsWith(matrix.config.os, 'ubuntu')
