### 2022-05-21 292a14f
Updating docs to remove unecessary animations to save memory
