### 2022-06-06 009e7c7
Correcting error for EU-12 EU-15 regions
