### 2022-06-08 7b1dbbe
Updating cheatsheet
