### 2022-06-08 78d959c
Updating cheatsheet
