+++ .github/workflows/docs.yaml
@@ -48,7 +48,7 @@ jobs:
           git config --local user.email "$GITHUB_ACTOR@users.noreply.github.com"
           Rscript -e 'pkgdown::deploy_to_branch(new_process = FALSE)'
           
-            - name: Track file
+      - name: Track file
         run: |
           git add .github/r-depends.rds
           git commit -m "track file"
