### 2022-06-09 #4015 @maczokni
Hi, I'm having trouble downloading the new version:

```
> devtools::install_github("JGCRI/rmap")

Downloading GitHub repo JGCRI/rmap@HEAD
Error in utils::download.file(url, path, method = method, quiet = quiet,  :
  download from 'https://api.github.com/repos/JGCRI/rmap/tarball/HEAD' failed
```
I also tried with remotes instead of devtools but no luck. I can download other repos no problem with the same method (e.g. `devtools::install_github("mpjashby/sfhotspot")` downloads the package no problem).

Here's my session info:

```
R version 4.2.0 (2022-04-22)
Platform: x86_64-apple-darwin17.0 (64-bit)
Running under: macOS Monterey 12.3.1

Matrix products: default
LAPACK: /Library/Frameworks/R.framework/Versions/4.2/Resources/lib/libRlapack.dylib

locale:
[1] en_GB.UTF-8/en_GB.UTF-8/en_GB.UTF-8/C/en_GB.UTF-8/en_GB.UTF-8

attached base packages:

 [1] rstudioapi_0.13   magrittr_2.0.3    usethis_2.1.6     devtools_2.4.3    pkgload_1.2.4     R6_2.5.1          rlang_1.0.2
 [8] fastmap_1.1.0     tools_4.2.0       pkgbuild_1.3.1    sessioninfo_1.2.2 cli_3.3.0         withr_2.5.0       ellipsis_0.3.2
[15] remotes_2.4.2     rprojroot_2.0.3   lifecycle_1.0.1   crayon_1.5.1      brio_1.1.3        processx_3.5.3    purrr_0.3.4
[22] callr_3.7.0       fs_1.5.2          ps_1.7.0          curl_4.3.2        testthat_3.1.4    memoise_2.0.1     glue_1.6.2
[29] cachem_1.0.6      compiler_4.2.0    desc_1.4.1        prettyunits_1.1.1
```

I suspect it will be to do with the package size and therefore time it takes to download. If I download via browser the tarball from the URL, the file is 120.9 (!!!) MB. Please see here about size and limitations for packages on cran: [https://thecoatlessprofessor.com/programming/r/size-and-limitations-of-packages-on-cran/](https://thecoatlessprofessor.com/programming/r/size-and-limitations-of-packages-on-cran/)

I am happy to continue the review having installed from the package archive file I've downloaded, but it might be something to consider to reduce the size (again referring to the cran guidance).
