### 2022-06-11 #4015 @maczokni
Thanks for this! It might be that I initially downloaded at work with better internet access. I am not an expert in packaging data into packages, but is there not a solution to appeal to a database? The package sizes for rnaturalearth and rnaturalearthdata are both much smaller (<1mb and 3mb) and from what I understand perform roughly the same functionality of finding a relevant shapefile from a query? It might introduce a separate step I guess (e.g. find all which meet some higher-level category e.g. ask user to specify country first, get those relevant geometries, and then match on the relevant "subRegion" column?). Anyway like I said this is not my area of expertise, so I'm not sure what solution to suggest.

Another thing: when you summon the help file for the package it just links to the github repo and github pages. The link for the cheatsheet is broken (it says https://github.com/JGCRI/rmap/blob/main/rmapCheatsheet.pdf instead of https://jgcri.github.io/rmap/cheatsheet.pdf)

Overall, I remain unconvinced that this is easy for those who are new to GIS, or easier than existing solutions. I don't think (at least to me) it is clear enough from the documentation how to go about applying this to my own data, which I believe is meant to be the strength of this package. I say this after running into more issues, this time trying to evoke the automatic "Multi-Year-Class-Scenario-Facet". The example in the documentation produces a dataframe with repeat values in the subRegion column, but when I try with the gcse data we used above I get an error about this (presumably when the map is being summoned):

```
library(rmap)
library(tidyverse)
library(readxl)
library(janitor)
library(sf)

# read in new test data
download.file("https://data.london.gov.uk/download/gcse-results-by-borough/a8a71d73-cc48-4b30-9eb5-c5f605bc845c/gcse-results.xlsx",
              destfile = "gcse-results.xlsx")
gcse_results <- read_xlsx("gcse-results.xlsx", sheet = "2020-21")
gcse_results_2 <- read_xlsx("gcse-results.xlsx", sheet = "2019-20")

# clean up test data
colnames <- paste0(gcse_results[1,], gcse_results[2,])
colnames <- gsub("NA", "", colnames)
names(gcse_results) <- colnames
gcse_results <- gcse_results %>%
  clean_names() %>%
  slice(4:36) %>%
  mutate(number_of_pupils_at_the_end_of_key_stage_4 = as.numeric(number_of_pupils_at_the_end_of_key_stage_4))

colnames <- paste0(gcse_results_2[1,], gcse_results_2[2,])
colnames <- gsub("NA", "", colnames)
names(gcse_results_2) <- colnames
gcse_results_2 <- gcse_results_2 %>%
  clean_names() %>%
  slice(4:36) %>%
  mutate(number_of_pupils_at_the_end_of_key_stage_4 = as.numeric(number_of_pupils_at_the_end_of_key_stage_4))

gcse_results_joined <- rbind(gcse_results %>% mutate(year = 2020),
                          gcse_results_2 %>% mutate(year = 2019))

rmap::map(data = gcse_results_joined,
           subRegCol = "area",
           valueCol = "number_of_pupils_at_the_end_of_key_stage_4",
           ncol=4,
           background = T)

Starting map...
Map data table written to /Users/user/Desktop/teaching/crime_mapping_discussions/outputs//map_param.csv
Error in rmap::map(data = gcse_results_joined, subRegCol = "area", valueCol = "number_of_pupils_at_the_end_of_key_stage_4",  :
  Input data data has multiple values. Please check your data.

```

I assumed we could get around this by extracting the map first, and using this. To do this try map_find() function. This function has not been changed like the rmap() function, and so still requires a renaming of the column to subRegion:

```
> map_find(gcse_results_joined, subRegCol = "area")
Error in map_find(gcse_results_joined, subRegCol = "area") :
  unused argument (subRegCol = "area")
> map_find(gcse_results_joined)
Error in map_find(gcse_results_joined) : data must have subRegion columns.

```

But even after creating a "subRegion" column in my data, this map doesn't help me:

```
gcse_results_joined$subRegion <- gcse_results_joined$area

thing <- map_find(gcse_results_joined)

rmap::map(data = gcse_results_joined,
          underLayer = thing
          background = T)

```

Specifying the value column summons the same error:

```
> rmap::map(data = gcse_results_joined,
+           underLayer = thing,
+           background = T,
+           valueCol = "number_of_pupils_at_the_end_of_key_stage_4")
Starting map...
Map data table written to /Users/user/Desktop/teaching/crime_mapping_discussions/outputs//map_param.csv
Error in rmap::map(data = gcse_results_joined, underLayer = thing, background = T,  :
  Input data data has multiple values. Please check your data.

```

Various specifications didn't help. In the end the only solution was to rename my data in the image of the data specified in the example:

```
data = data.frame(subRegion = gcse_results_joined$area,
                  year = gcse_results_joined$year,
                  value = gcse_results_joined$number_of_pupils_at_the_end_of_key_stage_4)

rmap::map(data = data,
          ncol=4,
          background = T)

```

So I think I'm misunderstanding the process here, and I would say I'm pretty comfortable with GIS and R, having published, and actively teaching in this space, but feel quite lost following the documentation with my own data. It is possible that to someone from a different background these processes are more intuitive, perhaps there is a need for this in the domain of the authors, where people have a certain approach/ understanding which is required to use this package. In that case I think this should be specified in the documentation.

Basically the behaviour I would expect would be something like:

```

rmap::map(data = gcse_results_joined,
          subRegion = area,
          year = year,
          value = number_of_pupils_at_the_end_of_key_stage_4,
          ncol=4,
          background = T)

```

My comment once again is that if the audience is people who have no GIS knowledge, and for some reason don't want to use rnaturalearth+tmap solutions to map their data, then:

1 - all functions should allow to specify where the relevant columns are in the own data (e.g. subRegion = "col", etc)
2 - the documentation should be improved with some clear instruction of how to apply the functions to an external data sets.
