+++ man/rmap.Rd
@@ -5,8 +5,11 @@
 \alias{rmap}
 \title{rmap}
 \description{
+`rmap` is a comprehensive mapping package to create easily customizable maps and compare across scenarios, years and data classes. The package allows creation of difference maps as well as customizable legends, color palettes and styles.
+}
+\details{
 \itemize{
 \item Github: https://github.com/JGCRI/rmap
 \item Webpage: https://jgcri.github.io/rmap/
-\item Cheatsheet: https://github.com/JGCRI/rmap/blob/master/rmapCheatsheet.pdf}
+\item Cheatsheet: https://jgcri.github.io/rmap/cheatsheet.pdf}
 }
