### 2022-06-22 4b62e5f
Updating gridded figures in docs
