+++ vignettes/vignette_map.Rmd
@@ -731,7 +731,9 @@ rmap::map(data)
 
 # User can specify that they want to plot this data on rmap::mapGCAMReg32
 rmap::map(data,
-          shape = rmap::mapGCAMReg32)
+          shape = rmap::mapGCAMReg32,
+          crop = F # Because will crop to the shapes boundaries when shape is specified
+          )
 
 ```
 
@@ -974,7 +976,7 @@ rmap::map(data = data,
 
 ```
 
-```{r, results = 'hide', eval=T, echo=F, warning=FALSE, error = FALSE, message = FALSE, fig.dim=c(15,15)}
+```{r, results = 'hide', eval=T, echo=F, warning=FALSE, error = FALSE, message = FALSE, fig.dim=c(15,5)}
 library(rmap);
 data = example_gridData_GWPv4To2015;
 mapx <- rmap::map(data = data,
@@ -1061,7 +1063,7 @@ mapx$map_param_KMEANS_xDiffAbs
 mapx$map_param_KMEANS_xDiffPrcnt
 ```
 
-```{r, results = 'hide', eval=T, echo=F, warning=FALSE, error = FALSE, message = FALSE, fig.dim=c(15,15)}
+```{r, results = 'hide', eval=T, echo=F, warning=FALSE, error = FALSE, message = FALSE, fig.dim=c(15,5)}
 library(rmap);
 
 data = example_gridData_GWPv4To2015
@@ -1107,7 +1109,7 @@ rmap::map(data = data,
 
 ### Gridded Data Multi-Class
 
-```{r, results = 'hide', eval=T, echo=TRUE, warning=FALSE, error = FALSE, message = FALSE, fig.dim=c(15,15)}
+```{r, results = 'hide', eval=T, echo=TRUE, warning=FALSE, error = FALSE, message = FALSE, fig.dim=c(15,5)}
 library(rmap);library(dplyr)
 
 # Change the years in the gridded data to classes
@@ -1262,7 +1264,7 @@ mapx$map_param_KMEANS
 
 ```
 
-```{r, results = 'hide', eval=T, echo=F, warning=FALSE, error = FALSE, message = FALSE, fig.dim=c(15,15)}
+```{r, results = 'hide', eval=T, echo=F, warning=FALSE, error = FALSE, message = FALSE, fig.dim=c(15,5)}
 library(rmap);
 
 data = example_gridData_GWPv4To2015 %>%
@@ -1366,7 +1368,7 @@ mapx$map_param_KMEANS_DiffPrcnt
 
 ```
 
-```{r, results = 'hide', eval=F, echo=F, warning=FALSE, error = FALSE, message = FALSE, fig.dim=c(15,15)}
+```{r, results = 'hide', eval=F, echo=F, warning=FALSE, error = FALSE, message = FALSE, fig.dim=c(15,5)}
 library(rmap);
 
 # # Change the years in the example gridded data to scenarios
@@ -1430,7 +1432,7 @@ rmap::map(data = data,
 
 ### Gridded Data Multi-Facet
 
-```{r, results = 'hide', eval=T, echo=TRUE, warning=FALSE, error = FALSE, message = FALSE, fig.dim=c(15,15)}
+```{r, results = 'hide', eval=T, echo=TRUE, warning=FALSE, error = FALSE, message = FALSE, fig.dim=c(15,5)}
 library(rmap);library(dplyr)
 
 # Create data with new columns RCPs and GCMs
