### 2022-06-22 3e08103
Updating cmdcheck with no vignettes
