+++ .github/workflows/build.yml
@@ -42,5 +42,5 @@ jobs:
         shell: Rscript {0}
            
       - name: Check
-        run: rcmdcheck::rcmdcheck(args = "--no-manual", error_on = "error")
+        run: rcmdcheck::rcmdcheck(args = c("--no-manual", "--ignore-vignettes", "--no-build-vignettes"), build_args = c("--no-manual", "--ignore-vignettes", "--no-build-vignettes"), error_on = "error")
         shell: Rscript {0}
