+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.4
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2022-06-22T16:59Z
+last_built: 2022-06-22T18:42Z
 
