### 2022-07-15 #4015 @hugoledoux
thanks @maczokni for the update, this is clear and I agree with what you wrote.

Reviewers are expected indeed to test the package, but reviewers are indeed not debuggers. That being said, there are often bugs in software and one can't expect all software to be perfect.

@CamilleMorlighem a point was raised about testing, and since you did the review a while ago (before changes to the package were made) could you do them again perhaps?
