### 2022-07-16 #4015 @zarrarkhan
@hugoledoux @maczokni

Thank you for your latest comments and the time you put into your thorough review. Your comments are addressed below and in the latest version of the paper draft: https://res.cloudinary.com/hkvhhndao/image/upload/v1657942383/ew2vgphg2dfqi6qy5n0t.pdf which can also be seen at:
https://github.com/JGCRI/rmap/blob/main/paper/paper.md

------------------------------------------------------

**Comment 1: If you must keep the table to map part, for the point "State of the field: Do the authors describe how this software compares to other commonly-used packages?" I would like to see discussion v rnaturalearth and other approaches to detail in what way this compares. I would like to be convinced in the paper that there is an added value here.**

**Response:** The `table to map` is one of several functionalities of `rmap` that makes it easy to use. We agree that it is a minor improvement and so remove it from the statement of need section as its own paragraph. We leave discussion of this functionality in the paper in other sections.

---------------------------------------------------------

**My other sticking point was "Substantial scholarly effort: Does this submission meet the scope eligibility described in the [JOSS guidelines](https://joss.readthedocs.io/en/latest/submitting.html#substantial-scholarly-effort)" specifically the point: "enables some new research challenges to be addressed or makes addressing research challenges significantly better (e.g., faster, easier, simpler)." The data table to map being faster/easier/simpler was something I was not convinced about, but with the long back and forth we are at a point where I no longer want to be carrying out extensive testing on behalf of the authors. On the other hand if the key contribution is now this ease for the comparison around the different scenarios, this should be emphasised more.**

**Response:**  Yes, one of the key contributions is the comparison around different scenarios and time periods. This is already emphasized in the paper as follows:

**- In the title of the paper:** "rmap: An R package to plot and compare tabular data on customizable maps across scenarios and time".
**- Abstract (lines 13 to 14):** Additionally rmap automatically detects and produces comparison maps if the data has multiple scenarios or time periods as well as animations for time series data
**- Statement of Need (lines 54 to 59):** Difference maps: Existing packages do not produce difference maps to compare across scenarios or time periods. rmap provides this functionality by automatically recognizing multiple scenarios and time periods to produce difference maps across these dimensions. An important aspect of spatial data is exploring the difference between two scenarios or time periods and rmap makes this a seamless process.
**- And as a Separate Section called Compare Scenarios (lines 84 -100):** With example code and figures.

Spatial comparisons across scenarios and time are very valuable in multiple fields and this is one of the features of `rmap` that has already been significantly used in the following published papers (and also in several additional papers in-progress and submission):
- Wild, T.B., Khan, Z., Clarke, L., Hejazi, M., Bereslawski, J.L., Suriano, M., Roberts, P., Casado, J., Miralles-Wilhelm, F., Gavino-Novillo, M. and Muñoz-Castillo, R., 2021. Integrated energy-water-land nexus planning in the Colorado River Basin (Argentina). Regional Environmental Change, 21(3), pp.1-16.
- Wild, T.B., Khan, Z., Zhao, M., Suriano, M., Bereslawski, J.L., Roberts, P., Casado, J., Gaviño‐Novillo, M., Clarke, L., Hejazi, M. and Miralles‐Wilhelm, F., 2021. The Implications of Global Change for the Co‐Evolution of Argentina's Integrated Energy‐Water‐Land Systems. Earth's Future, 9(8), p.e2020EF001970.
- Sampedro, J., Khan, Z., Vernon, C.R., Smith, S.J., Waldhoff, S. and Van Dingenen, R., 2022. rfasst: An R tool to estimate air pollution impacts on health and agriculture. Journal of Open Source Software, 7(69), p.3820.
- Khan, Z., Wild, T.B., Iyer, G., Hejazi, M. and Vernon, C.R., 2021. The future evolution of energy-water-agriculture interconnectivity across the US. Environmental Research Letters, 16(6), p.065010.
- Khan, Z., Iyer, G., Patel, P., Kim, S., Hejazi, M., Burleyson, C. and Wise, M., 2021. Impacts of long-term temperature change and variability on electricity investments. Nature communications, 12(1), pp.1-12.
- Khan, Z., Wild, T.B., Carrazzone, M.E.S., Gaudioso, R., Mascari, M.P., Bianchi, F., Weinstein, F., Pérez, F., Pérez, W., Miralles-Wilhelm, F. and Clarke, L., 2020. Integrated energy-water-land nexus planning to guide national policy: an example from Uruguay. Environmental Research Letters, 15(9), p.094014.

---------------------------------------------------------
