+++ pkgdown.yml
@@ -1,7 +1,7 @@
 pandoc: 2.7.3
-pkgdown: 2.0.4
+pkgdown: 2.0.5
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2022-06-22T18:42Z
+last_built: 2022-07-16T03:41Z
 
