### 2022-08-16 #4015 @CamilleMorlighem
Hi all,

Following comments from @hugoledoux and @maczokni, I have redone old tests of the package I did a while ago and new tests considering the updates of the package. See the following examples:

 ### Load libraries
```
library(stringr)
library(dplyr)
library(rmap)
```
 ### Example 1 : Senegalese population per region
```
# Get population data
download.file("https://data.humdata.org/dataset/c686be28-1858-44e9-a44b-598ce32ca20b/resource/85ba6660-fad2-4e49-b9b3-9ecb754d5700/download/sen_admpop_adm1_2020.csv",
              destfile = "sen_admpop_adm1_2020.csv", mode ="wb")
reg_pop = read.csv("sen_admpop_adm1_2020.csv")

# Clean data : missing accent on the "e"
reg_pop = reg_pop %>%
    mutate(ADM1_FR = str_replace_all(ADM1_FR, c("Thies" = "Thiès", "Kedougou" = "Kédougou", "Sedhiou" = "Sédhiou")))

map(reg_pop,
    subRegion = "ADM1_FR",
    value = "Total",
    legendType = "pretty",
    title = "Population per region in Senegal")
```
Note that the region name of three regions is misspelt (Thiès, Kédougou, Sédhiou) in the downloaded data with missing acute and grave accents, although the `ADM1_FR` column should contain the French names of the regions (and hence with accents).

![map_param_PRETTY](https://user-images.githubusercontent.com/85868805/184907713-6f8893fa-3b37-4ca8-92be-5ca5fb7014f7.png)

 ### Example 2 : Covid-19 cases per African country
```
# Get covid-19 data
download.file("https://data.humdata.org/dataset/9151988e-ae09-416c-9408-6df93bebf3a2/resource/ab898c39-2322-4bf0-a9a7-5897f5605bc1/download/africa_covid19_daily_deaths_national.csv",
              destfile = "africa_cov_cases.csv", mode ="wb")
cases_africa = read.csv("africa_cov_cases.csv", sep = ";")

# Clean data : aggregate cases, remove empty rows and correct region names
cases_per_country = cases_africa %>%
    mutate(sum = rowSums(cases_africa[,4:length(cases_africa)], na.rm = T)) %>%
    subset(ISO != "") %>%
    mutate(COUNTRY_NAME = str_replace_all(COUNTRY_NAME, c("Sao Tome and Principe" = "São Tomé and Principe", "Eswatini" = "eSwatini")))

map(cases_per_country,
    subRegion = "COUNTRY_NAME",
    value = "sum",
    title = "Covid-19 cases per African country")
```
Note here that the names of eSwatini and São Tomé and Principe were misspelt in the downloaded data (respectively Eswatini and Sao Tome and Principe).

![map_param_KMEANS](https://user-images.githubusercontent.com/85868805/184908142-d2b9fcda-3547-4ca7-bf00-79f79019ddf2.png)
 ### Example 3 : Infectious diseases in California
#### Plot different years
```
# Get infectious disease data
download.file("https://data.chhs.ca.gov/dataset/03e61434-7db8-4a53-a3e2-1d4d36d6848d/resource/75019f89-b349-4d5e-825d-8b5960fc028c/download/odp_idb_2020_ddg_compliant.csv",
              destfile = "infect_disease_us_county.csv", mode ="wb")

i_disease = read.csv("infect_disease_us_county.csv") ; colnames(i_disease)

# Clean data : keep years 2018-2020 and cases of Malaria, Typhus, Dengue and Lyme diseases
i_4_disease <- i_disease %>%
    subset(grepl("Malaria|Typhus|Dengue|Lyme", Disease) & grepl("Total", Sex) & grepl("2018|2019|2020", Year)) %>%
    filter(County != "California")  %>% # csv contains a row with sum of cases for all counties in California
    mutate(County = paste0(County, "_CA")) # add proper state abbreviation to match built-in map counties

malaria_map = map(i_4_disease[i_4_disease == "Malaria", ],
                  subRegion = "County",
                  value = "Cases",
                  save = F,
                  x = "Year",
                  title = "Malaria cases")

malaria_map$`map_param_KMEANS`
```
![map_param_KMEANS](https://user-images.githubusercontent.com/85868805/184909670-8dd7a2a4-be4b-431e-bdb5-06a66e512ab6.png)

#### Plot different diseases
```
disease_map = map(i_4_disease[i_4_disease$Year == "2020", ],
                  subRegion = "County",
                  value = "Cases",
                  save = F,
                  class = "Disease",
                  ncol = 4,
                  title = "Cases of several infectious diseases")

disease_map$`map_param_KMEANS`
```
![map_param_KMEANS](https://user-images.githubusercontent.com/85868805/184909895-bb1e091b-6b80-4056-b0b9-c11b2181f5d3.png)

#### Compare across years
```
malaria_map_diff_2018 = map(i_4_disease[i_4_disease == "Malaria", ],
                            subRegion = "County",
                            value = "Cases",
                            save = F,
                            xRef = "2018",
                            underLayer = mapUS49County,
                            title = "Difference with 2018 malaria cases")

malaria_map_diff_2018$map_param_KMEANS_xDiffAbs
malaria_map_diff_2018$map_param_KMEANS_xDiffPrcnt
```
![map_param_KMEANS_xDiffAbs](https://user-images.githubusercontent.com/85868805/184910951-e791b8e0-9f23-4600-89b8-8b096c3f0acf.png)
![map_param_KMEANS_xDiffPrcnt](https://user-images.githubusercontent.com/85868805/184910967-dcf9ed44-0964-4a88-a15b-228ce16159d9.png)

For some reason, `malaria_map_diff_2018$map_param_KMEANS_xDiffPrcnt` does not display values for some counties. I think this might be because extreme values of the legend classes are not comprised in any classes ? (e.g. Counties where exactly 0% change occurs are not represented in any class). I think this issue needs to be resolved. Adding square brackets around the extreme values of the classes could help specify in which class the extreme values are included.

After retesting the package, from my point of view, I think it is the responsibility of the user to make sure its data are complete, up-to-date and correct (e.g. make sure all Region names are properly written) and match the “subregions” of the `rmap` built-in maps. In any case, if the user had to use another solution than `rmap`, and join its data for example with the attribute table of an existing shapefile, he/she would still need to clean its data beforehand. Following @maczokni comments, maybe a section of the user guide could highlight the need for cleaning data and matching them with the built-in maps subRegions?

As for the novelty of the package, as raised by @maczokni, I think its true novelty lies indeed in the ability to easily compare across year/scenarios/categories. On another note, compared to existing packages like `rnaturalearth` and `tigris`, everything can be done here using only the package `rmap`, while the two former solutions require combining it with other packages such as `tmap`.  Another advantage of `rmap` (that is maybe not highlighted enough in the paper) is that it creates `ggplot` objects which can easily be combined with others to change the map layout, for example, and I think `ggplot` might be more known among non-GIS users than `tmap`.

Except for the issue with the legend classes, I think `rmap` worked as expected in the examples above. Let me know what you think @hugoledoux and @maczokni.
