### 2022-08-17 #4015 @hugoledoux
okay, I re-read the (very detailed and good) comments from the reviewers (thanks @CamilleMorlighem and @maczokni) and I think we can go ahead and try to finalise the submission.

I would like the 3 points raised above to be fixed:

1. add a section about how to clean the data and match them with the maps subregions; this doesn't need to be in the paper but in the docs it would be necessary I believe, both reviewers struggle with this
2. extreme values not being displayed for some datasets: can you look into that?
3. in the "Statement of Need", perhaps add what @CamilleMorlighem mentions about `ggplot` objects?

Once this is done, I am happy to process further with acceptance of the submission
