### 2022-09-07 557b906
Updating rmap in response to https://github.com/openjournals/joss-reviews/issues/4015#issuecomment-1216784122
