### 2022-09-07 e53de5f
Adding section on showNA
