### 2022-09-07 7ef7413
setting showNA to T
