+++ R/map.R
@@ -179,7 +179,7 @@ map <- function(data = NULL,
                 palette = NULL,
                 paletteDiff = "pal_div_BluRd",
                 colorNA = "gray50",
-                showNA = F,
+                showNA = T,
                 ncol = 3,
                 size = 16,
                 alpha = 1,
