### 2022-09-07 1a6ceda
setting showNA to F
