### 2022-09-07 35f442f
Updating docs to avoid dplyr error
