+++ pkgdown.yml
@@ -1,7 +1,7 @@
 pandoc: 2.7.3
-pkgdown: 2.0.5
+pkgdown: 2.0.6
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2022-07-16T03:41Z
+last_built: 2022-09-07T20:04Z
 
