+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.6
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2022-09-07T20:04Z
+last_built: 2022-09-07T22:05Z
 
