### 2022-09-08 #4015 @zarrarkhan
@hugoledoux

Please see each of the points mentioned above addressed in the latest version of the paper.

1. add a section about how to clean the data and match them with the maps subregions; this doesn't need to be in the paper but in the docs it would be necessary I believe, both reviewers struggle with this

Section on [Cleaning Data](https://jgcri.github.io/rmap/articles/vignette_map.html#clean) added to user guide explaining how subRegions will need to be matched to built-in map names along with an example of how rmap automatically identifies the missing regions for you and how you can then go in and find the correct spelling.

2. extreme values not being displayed for some datasets: can you look into that?

@CamilleMorlighem Thanks for catching this. The issue here actually was that the % difference calculation uses the base year value to calculate the % difference i.e. Percent Diff = (Value in Year_i - Value in Year_Ref)*100/(Value in Year_Ref). In the examples you provided or in general when there was a 0 value in the Ref year the result was coming out to be Infinite ('Inf'). This was then resulting in an NaN value being filtered out and thus those counties being removed. We have fixed this to behave better as follows:

- If there is a '0' value in the Ref year `rmap` gives a warning saying that there is a '0' value and the difference calculation will give an `Inf` value which is converted to an `NA`.
- `rmap` comes with the option of highlight single values such as `0`s as well as showing NA values (listed in separate sections of the user guide: [Legend Single Value](https://jgcri.github.io/rmap/articles/vignette_map.html#legend-single-value) and [Show NA](https://jgcri.github.io/rmap/articles/vignette_map.html#showna)
- We switched the default `showNA` value to `T` so it shows these.

If we redo your example with `legendSingleValue = 0`  and `showNA=0` you get the following.
This time rmap gives the warning about the Inf values and we can see "NAs" vs. '0' values in the maps.

```
library(stringr)
library(dplyr)
library(rmap)

# Get infectious disease data
download.file("https://data.chhs.ca.gov/dataset/03e61434-7db8-4a53-a3e2-1d4d36d6848d/resource/75019f89-b349-4d5e-825d-8b5960fc028c/download/odp_idb_2020_ddg_compliant.csv",
              destfile = "infect_disease_us_county.csv", mode ="wb")

i_disease = read.csv("infect_disease_us_county.csv") ; colnames(i_disease)

# Clean data : keep years 2018-2020 and cases of Malaria, Typhus, Dengue and Lyme diseases
i_4_disease <- i_disease %>%
  subset(grepl("Malaria|Typhus|Dengue|Lyme", Disease) & grepl("Total", Sex) & grepl("2018|2019|2020", Year)) %>%
  filter(County != "California")  %>% # csv contains a row with sum of cases for all counties in California
  mutate(County = paste0(County, "_CA")) # add proper state abbreviation to match built-in map counties

malaria_map_diff_2018 = map(i_4_disease[i_4_disease == "Malaria", ],
                            subRegion = "County",
                            value = "Cases",
                            save = F, show=F,
                            xRef = "2018",
                            underLayer = mapUS49County,
                            labels=T, legendSingleValue = 0)

malaria_map_diff_2018$map_param_KMEANS
malaria_map_diff_2018$map_param_KMEANS_xDiffAbs
malaria_map_diff_2018$map_param_KMEANS_xDiffPrcnt
```
![map_param_KMEANS](https://user-images.githubusercontent.com/7863115/188962335-81cb05fb-a659-4917-a0a1-25dc8588c26c.png)
![map_param_KMEANS_xDiffAbs](https://user-images.githubusercontent.com/7863115/188962344-ce630dbd-a1bb-4e6a-841a-e0093c32d983.png)
![map_param_KMEANS_xDiffPrcnt](https://user-images.githubusercontent.com/7863115/188962355-13017304-6388-426a-84f5-72cad3fba290.png)

3. in the "Statement of Need", perhaps add what @CamilleMorlighem mentions about ggplot objects?

We mention the point about ggplot objects in the Statement of Needs as follows:

"**Post-process customization**: Existing packages do not produce output objects that can be saved and then customized. Customization of the maps in existing packages is limited to package specific functionality and arguments. `rmap` produces `ggplot` objects in which every element (axis, grids, titles, colors, line widths, facets) can all be customized after the map has been produced. This allows users to capitalize on existing knowledge of the widely used `ggplot2` package and its arguments."
