### 2022-09-12 #4015 @hugoledoux
@zarrarkhan thanks, everything looks good in the paper and the docs.

I have 2 small comments/questions:

1. what are ggplot "arguments"? (l.53+56). I don't use ggplot and I wouldn't know what you're referring to, perhaps just use "functionalities"?
2. the reference for Pebesma simple feature has the name of the journal missing, and then the github repo is cited. To fix.
