+++ paper/paper.bib
@@ -112,9 +112,9 @@
 	title = {Simple features for {R}: standardized support for spatial vector data.},
 	volume = {10},
 	shorttitle = {Simple features for {R}},
-	url = {https://github.com/r-spatial/sf},
+	url = {https://journal.r-project.org/archive/2018/RJ-2018-009/index.html},
 	number = {1},
-	journal = {R J.},
+	journal = {The R Journal},
 	author = {Pebesma, Edzer},
 	year = {2018},
 	pages = {439},
