### 2022-09-12 #4015 @zarrarkhan
> @zarrarkhan thanks, everything looks good in the paper and the docs.
>
> I have 2 small comments/questions:
>
> 1. what are ggplot "arguments"? (l.53+56). I don't use ggplot and I wouldn't know what you're referring to, perhaps just use "functionalities"?
> 2. the reference for Pebesma simple feature has the name of the journal missing, and then the github repo is cited. To fix.

Thanks @hugoledoux. Both comments addressed as follows in latest push:
http://res.cloudinary.com/hkvhhndao/image/upload/v1663018510/xpwhvwopcw7c7jm9q3ir.pdf

1. Updated to say ggplot `functionality` and `ggplot chart objects`
2. Reference fixed with full journal name added (The R Journal).

Let me know if you have additional comments/edits for me to make.
Best
Zarrar
