### 2022-09-14 #4015 @zarrarkhan
@hugoledoux

Addressed as follows:
- Tagged version of software: v1.0.2 https://github.com/JGCRI/rmap/releases/tag/v1.0.2
- Archived on Zenodo: https://doi.org/10.5281/zenodo.7080178
- Metadata checked with correct title matching paper title and all author names with ORCID IDs from paper.
- DOI: 10.5281/zenodo.7080178

Please let me know if you need anything else.

Best

Zarrar
