+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.6
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2022-09-12T21:44Z
+last_built: 2022-09-14T20:26Z
 
