### 2022-09-15 #4015 @zarrarkhan
@hugoledoux

oops, sorry about that. Thanks for catching that! Updated and minted again:

- Tagged version of software: v1.0.3 https://github.com/JGCRI/rmap/releases/tag/v1.0.3
- Archived on Zenodo: https://zenodo.org/record/7082820#.YyNRrXbMJaQ
- Metadata checked with correct title matching paper title and all author names with ORCID IDs from paper.
- DOI: 10.5281/zenodo.7082820

Best

Zarrar
