+++ pkgdown.yml
@@ -3,5 +3,5 @@ pkgdown: 2.0.6
 pkgdown_sha: ~
 articles:
   vignette_map: vignette_map.html
-last_built: 2022-09-15T15:07Z
+last_built: 2022-09-15T16:28Z
 
