### 2022-09-15 8e53cc8
Updating map_plot to not atrificially change data subRegion names.
