+++ R/map_plot.R
@@ -389,26 +389,14 @@ if(T){ # Initialize
   # Rename certain countries to rmap names
   if("subRegion" %in% names(data)){
     data <- data %>%
-      dplyr::mutate(subRegion = dplyr::case_when(grepl("^United States of America$|^United States$",subRegion,ignore.case = T)~"USA",
-                                                 grepl("^Tanzania$",subRegion,ignore.case = T)~"United Republic of Tanzania",
-                                                 grepl("^Democratic Republic of Congo$",subRegion,ignore.case = T)~"Democratic Republic of the Congo",
-                                                 grepl("^Congo$",subRegion,ignore.case = T)~"Republic of the Congo",
-                                                 grepl("^Cote d'Ivoire$",subRegion,ignore.case = T)~"Ivory Coast",
-                                                 grepl("^Serbia$",subRegion,ignore.case = T)~"Republic of Serbia",
-                                                 grepl("^EU-12$",subRegion,ignore.case = T)~"EU_12",
+      dplyr::mutate(subRegion = dplyr::case_when(grepl("^EU-12$",subRegion,ignore.case = T)~"EU_12",
                                                  grepl("^EU-15$",subRegion,ignore.case = T)~"EU_15",
                                                  TRUE ~ subRegion))
   }
 
   if("region" %in% names(data)){
     data <- data %>%
-      dplyr::mutate(region = dplyr::case_when(grepl("^United States of America$|^United States$",region,ignore.case = T)~"USA",
-                                              grepl("^Tanzania$",region,ignore.case = T)~"United Republic of Tanzania",
-                                              grepl("^Democratic Republic of Congo$",region,ignore.case = T)~"Democratic Republic of the Congo",
-                                              grepl("^Congo$",region,ignore.case = T)~"Republic of the Congo",
-                                              grepl("^Cote d'Ivoire$",region,ignore.case = T)~"Ivory Coast",
-                                              grepl("^Serbia$",region,ignore.case = T)~"Republic of Serbia",
-                                              grepl("^EU-12$",region,ignore.case = T)~"EU_12",
+      dplyr::mutate(region = dplyr::case_when(grepl("^EU-12$",region,ignore.case = T)~"EU_12",
                                               grepl("^EU-15$",region,ignore.case = T)~"EU_15",
                                               TRUE ~ region))
   }
