### 2022-09-16 #4015 @danielskatz
Hi @zarrarkhan - I'm the AEiC on duty this week, and have proofread your paper. Please merge https://github.com/JGCRI/rmap/pull/92 for some very minor changes in it.

It appears https://doi.org/10.1002/wics.147 is a DOI for the Wickham reference. If you agree, please add it. Similar, I think at least a URL is needed for Brunsdon & Chen.

Once these are done, we can move to acceptance and publication.
