### 2022-09-16 f7309c2
Updating paper and bib
