+++ paper/paper.bib
@@ -72,13 +72,6 @@
 	year = {2011}
 }
 
-@article{brunsdon_package_2015,
-	title = {Package ‘{GISTools}’},
-	journal = {CRAN Repository},
-	author = {Brunsdon, Chris and Chen, Hongyan},
-	year = {2015}
-}
-
 @misc{qgis_development_team_qgis_2021,
 	title = {{QGIS}},
 	url = {qgis.org/en/site/},
@@ -146,6 +139,7 @@
 	title = {ggplot2},
 	volume = {3},
 	number = {2},
+	doi = { https://doi.org/10.1002/wics.147},
 	journal = {Wiley Interdisciplinary Reviews: Computational Statistics},
 	author = {Wickham, Hadley},
 	year = {2011},
