### 2022-09-16 #4015 @zarrarkhan
@danielskatz @hugoledoux

Thank you so much for all your comments and edits. We agree that the review process has been very helpful and we feel much more confident in the usefulness of the software for multiple users. Updated and released in v1.0.4 with following changes:

- Wheadon preview: http://res.cloudinary.com/hkvhhndao/image/upload/v1663333163/oj1b28zu3aml6r9nb4zx.pdf
- DOI for the Wickham reference added: https://doi.org/10.1002/wics.147
- Brunsdon & Chen reference removed since it is no longer available on CRAN and does not look like it is being maintained.

Details for latest release
- Tagged version of software: v1.0.5 https://github.com/JGCRI/rmap/releases/tag/v1.0.5
- Archived on Zenodo: https://zenodo.org/record/7085969#.YyR2HXbMJaQ
- Metadata checked with correct title matching paper title and all author names with ORCID IDs from paper.
- DOI: 10.5281/zenodo.7085969

Hope we caught everything!

Best

Zarrar
