### 2022-09-16 #4015 @danielskatz
@zarrarkhan - note that the new DOI should be `doi = {10.1002/wics.147},` instead of `doi = { https://doi.org/10.1002/wics.147},`

You do not need to generate a new version or a new archive for this, since it will only be a change in the paper, which is not required to be in the archived repo, but if you want to, you can.
