+++ paper/paper.bib
@@ -139,7 +139,7 @@
 	title = {ggplot2},
 	volume = {3},
 	number = {2},
-	doi = { https://doi.org/10.1002/wics.147},
+	doi = {10.1002/wics.147},
 	journal = {Wiley Interdisciplinary Reviews: Computational Statistics},
 	author = {Wickham, Hadley},
 	year = {2011},
