### 2022-09-16 #4015 @danielskatz
sorry @zarrarkhan - there is one more change needed - for the first Pebesma reference, please add `doi = {10.32614/RJ-2018-009},` rather than the URL that is there now.
