+++ paper/paper.bib
@@ -107,6 +107,7 @@
 	shorttitle = {Simple features for {R}},
 	url = {https://journal.r-project.org/archive/2018/RJ-2018-009/index.html},
 	number = {1},
+	doi = {10.32614/RJ-2018-009},
 	journal = {The R Journal},
 	author = {Pebesma, Edzer},
 	year = {2018},
