+++ paper/paper.bib
@@ -105,7 +105,6 @@
 	title = {Simple features for {R}: standardized support for spatial vector data.},
 	volume = {10},
 	shorttitle = {Simple features for {R}},
-	url = {https://journal.r-project.org/archive/2018/RJ-2018-009/index.html},
 	number = {1},
 	doi = {10.32614/RJ-2018-009},
 	journal = {The R Journal},
